% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DC_basisOfRecordBadlyFormed.R
\docType{data}
\name{basisOfRecordBadlyFormed}
\alias{basisOfRecordBadlyFormed}
\title{Data check basisOfRecordBadlyFormed Check if the specific nature of the data record is within controlled vocabulary}
\format{An object of class \code{"dataCheck"}, see \code{\link{dataCheck}} for details.}
\description{
This data check answers: "?" question.\cr Data check will pass if \strong{dwc:basisOfRecord was matched to vocabulary} and will fail if \strong{dwc:basisOfRecord could not be unambiguously matched to vocabulary}.\cr Dimension of this data check is \strong{Conformance} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{}, such data check would return \code{}.\cr Example of entries that will fail: \code{}, such data check would return \code{}.
}
\references{
LATER
}
