% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCclass_create.R
\name{createDCclassMain}
\alias{createDCclassMain}
\title{Create a data check object from a given YAML file}
\usage{
createDCclassMain(DCyaml)
}
\arguments{
\item{DCyaml}{Data check slot in YAML format}
}
\description{
Create a data check object from a given YAML file
}
