% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCclass_create.R
\name{createDCclassMeta}
\alias{createDCclassMeta}
\title{Create a data check metatadata from a given YAML file}
\usage{
createDCclassMeta(DCmeta)
}
\arguments{
\item{DCmeta}{Data check metadata slot in YAML format}
}
\description{
Create a data check metatadata from a given YAML file
}
