% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics_methods.R
\docType{class}
\name{DataCheck-class}
\alias{DataCheck-class}
\alias{DataCheck}
\alias{show,DataCheck-method}
\alias{dataCheck}
\title{Create Data Check Class}
\usage{
\S4method{show}{DataCheck}(object)
}
\arguments{
\item{object}{a dataCheck object}
}
\description{
Create Data Check Class

Show method for dataCheck objects
}
\section{Slots}{

\describe{
\item{\code{name}}{Official name of biodiversity data check}

\item{\code{input}}{Additional options to perform a data check}

\item{\code{description}}{Information about the data check}

\item{\code{flags}}{Flagging type for specific data check}

\item{\code{pseudocode}}{R pseudocode for a data check}

\item{\code{source}}{Information about creators and maintenance}
}}

