% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCclass_create.R, R/DCclass_perform.R
\docType{methods}
\name{show,dataCheck-method}
\alias{show,dataCheck-method}
\alias{show-dataCheck}
\alias{performDC}
\alias{performDC,dataCheck-method}
\title{Perform Data Checks}
\usage{
\S4method{show}{dataCheck}(object)

performDC(DC, DATA)

\S4method{performDC}{dataCheck}(DC, DATA)
}
\arguments{
\item{DC}{A data check as a dataCheck class object}

\item{DATA}{a data frame to perform data check on}
}
\value{
A vector of logical values that determine if data check was passed on 
specific entry in a given DATA object
}
\description{
Method that performs data check on a given dataset
}
