% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_export.R
\name{datacheck_info_export}
\alias{datacheck_info_export}
\title{Export data checks as R objects}
\usage{
datacheck_info_export(
  path_yaml = system.file("extdata/data_check.yaml", package = "bdchecks"),
  path_rd = "./R/",
  id_rd = "documentation_"
)
}
\arguments{
\item{path_yaml}{Path to a YAML file.}

\item{path_rd}{Path for .R (roxygen2) file.}

\item{id_rd}{ID in the beginning of file for .R file
(helps to differentiate other functions from data checks).}
}
\description{
`dc_export()` is a function for exporting data checks from YAML file to
rda and/or roxygen2 file.
}
