% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics_methods.R
\docType{methods}
\name{show,datacheck_single-method}
\alias{show,datacheck_single-method}
\alias{dataCheck}
\title{Show method for dataCheck objects}
\usage{
\S4method{show}{datacheck_single}(object)
}
\arguments{
\item{object}{a dataCheck object}
}
\description{
Show method for dataCheck objects
}
