% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCclass_create.R
\docType{class}
\name{dataCheckFlag-class}
\alias{dataCheckFlag-class}
\alias{dataCheckFlag}
\title{Combined data checks class}
\description{
Combined data checks class
}
\section{Slots}{

\describe{
\item{\code{DC}}{names of performed data checks}

\item{\code{flags}}{list of performed data checks in a dataCheckFlag_SINGLE class}

\item{\code{dataOrig}}{original data before data checks}

\item{\code{dataMod}}{modified data after data checks}
}}

