% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCclass_create.R
\docType{class}
\name{dataCheckFlag_SINGLE-class}
\alias{dataCheckFlag_SINGLE-class}
\alias{dataCheckFlag_SINGLE}
\title{Single data check flag class}
\description{
Single data check flag class
}
\section{Slots}{

\describe{
\item{\code{name}}{of performed data check}

\item{\code{target}}{column for performed data check}

\item{\code{flag}}{class}

\item{\code{result}}{logical vector}
}}

