% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics_methods.R
\docType{class}
\name{DataCheckFlagSet-class}
\alias{DataCheckFlagSet-class}
\alias{DataCheckFlagSet}
\alias{show,DataCheckFlagSet-method}
\title{Combined Data Checks Class}
\usage{
\S4method{show}{DataCheckFlagSet}(object)
}
\arguments{
\item{object}{a DataCheckFlagSet object}
}
\description{
Combined Data Checks Class

Show method for DataCheckFlagSet objects
}
\section{Slots}{

\describe{
\item{\code{DC}}{names of performed data checks}

\item{\code{flags}}{list of performed data checks in a DataCheckFlagSet class}

\item{\code{not_performed}}{data.frame of not performed data checks in a 
DataCheckFlagSet class}

\item{\code{flag_table}}{data.frame of original data with it's flagging information}
}}

