% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics_methods.R
\docType{class}
\name{DataCheckSet-class}
\alias{DataCheckSet-class}
\alias{DataCheckSet}
\title{Create Data Check Class}
\description{
Create Data Check Class
}
\section{Slots}{

\describe{
\item{\code{dc_body}}{Set of all available data checks}
}}

