% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DC_dayMonthTransposed.R
\docType{data}
\name{dayMonthTransposed}
\alias{dayMonthTransposed}
\title{Data check dayMonthTransposed Checks if day and month information is not transposed}
\format{An object of class \code{"dataCheck"}, see \code{\link{dataCheck}} for details.}
\description{
This data check answers: "Is day and month not switched?" question.\cr Data check will pass if \strong{The event month and day is feasible} and will fail if \strong{The event month is greater than 12 and event day is less than 12 inferring field transpositon}.\cr Dimension of this data check is \strong{Conformance} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{day=1 + month = 12, day = 31 + month = 1}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{day=1 + month = 13, day = 12 + month = 31}, such data check would return \code{Amendment}.
}
\references{
LATER
}
\keyword{date}
