% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_basisofrecord_present.R,
%   R/documentation_basisofrecord_present.R
\name{dc_basisofrecord_present}
\alias{dc_basisofrecord_present}
\title{Data check basisofrecord_present Check if basisOfRecord exists and is not empty.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_basisofrecord_present(TARGET = NULL)
}
\arguments{
\item{TARGET}{a vector of basisOfRecord information. To pass it must be
non-empty entry.}
}
\description{
This data check answers: "Is basisOfRecord information present?" question.\cr Data check will pass if \strong{The value for basisOfRecord is present.} and will fail if \strong{The value for basisOfRecord is missing.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{basisOfRecord=LivingSpecimen}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{basisOfRecord=}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for basisOfRecord is present.
}

\section{sampleFailData}{

The value for basisOfRecord is missing.
}

\section{targetDWCField}{

basisOfRecord
}

\section{checkCategory}{

 record-level terms
}

\examples{
perform_dc(data_bats, 'basisofrecord_present')
}
\references{
None
}
\keyword{completeness,}
\keyword{incomplete}
\keyword{missing,}
\keyword{other,}
\keyword{validation,}
