% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_basisofrecord_standard.R,
%   R/documentation_basisofrecord_standard.R
\name{dc_basisofrecord_standard}
\alias{dc_basisofrecord_standard}
\title{Data check basisofrecord_standard Check if basisOfRecord is specified source authority service.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_basisofrecord_standard(
  TARGET = NULL,
  sources = c("event", "fossil", "fossilspecimen", "human", "humanobservation",
    "livingspecimen", "machineobservation", "materialsample", "occurrence",
    "preservedspecimen", "photo", "photos", "taxon")
)
}
\arguments{
\item{TARGET}{a vector of basisOfRecord information. To pass it must be
within given reference dictionary.}

\item{sources}{a vector of available sources.}
}
\description{
This data check answers: "Is basisOfRecord within given source authority list?" question.\cr Data check will pass if \strong{The value for basisOfRecord is present and correct.} and will fail if \strong{The value for basisOfRecord is missing or incorrect.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{basisOfRecord=human_observation}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{basisOfRecord=photos}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for basisOfRecord is present and correct.
}

\section{sampleFailData}{

The value for basisOfRecord is missing or incorrect.
}

\section{targetDWCField}{

basisOfRecord
}

\section{checkCategory}{

 record-level terms
}

\examples{
perform_dc(data_bats, 'basisofrecord_standard')
}
\references{
None
}
\keyword{vocabulary,validation}
