% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_coordinatePrecisionMismatch.R,
%   R/documentation_coordinatePrecisionMismatch.R
\name{dc_coordinatePrecisionMismatch}
\alias{dc_coordinatePrecisionMismatch}
\title{Data check coordinatePrecisionMismatch}
\format{An object of class function to perform a specific data check.}
\usage{
dc_coordinatePrecisionMismatch(TARGET, DEPEND)
}
\arguments{
\item{TARGET}{a vector to perform data check}

\item{DEPEND}{a vector of coordinate precision value}
}
\description{
This data check answers: "?" question.\cr Data check will pass if \strong{} and will fail if \strong{}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{}, such data check would return \code{}.\cr Example of entries that will fail: \code{}, such data check would return \code{}.
}
\section{samplePassData}{

The number of decimal places of latitude (dwc:decimalLatitude) and longitude (dwc:decimalLongitude) is in agreement with the supplied coordinate precision (dwc:coordinatePrecision)
}

\section{sampleFailData}{

The number of decimal places of latitude (dwc:decimalLatitude) or longitude (dwc:decimalLongitude) is not in agreement with the supplied coordinate precision (dwc:coordinatePrecision)
}

\section{targetDWCField}{

decimalLatitude,decimalLongitude
}

\section{checkCategory}{

 Location
}

\examples{
perform_dc(data_bats, 'coordinatePrecisionMismatch')
}
\references{
None
}
