% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_coordinatesZero.R,
%   R/documentation_coordinatesZero.R
\name{dc_coordinatesZero}
\alias{dc_coordinatesZero}
\title{Data check coordinatesZero Check if decimal values given in decimalLatitude and decimalLongitude are not zero}
\format{An object of class function to perform a specific data check.}
\usage{
dc_coordinatesZero(TARGET)
}
\arguments{
\item{TARGET}{a vector of coordinate information}
}
\description{
This data check answers: "Is decimal number in langitude and latitude not zero?" question.\cr Data check will pass if \strong{Decimal latitude and longitude are not zero (0) degrees} and will fail if \strong{Decimal latitude and longitude are both zero (0) degrees}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{decimalLatitude=-1, decimalLongitude=1}, such data check would return \code{Passed}.\cr Example of entries that will fail: \code{decimalLatitude=0, decimalLongitude=0}, such data check would return \code{Failed}.
}
\section{samplePassData}{

Decimal latitude and longitude are not zero (0) degrees
}

\section{sampleFailData}{

Decimal latitude and longitude are both zero (0) degrees
}

\section{targetDWCField}{

decimalLatitude,decimalLongitude
}

\section{checkCategory}{

 Location
}

\examples{
perform_dc(data_bats, 'coordinatesZero')
}
\references{
None
}
