% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_coordinateuncertainty_inrange.R,
%   R/documentation_coordinateuncertainty_inrange.R
\name{dc_coordinateuncertainty_inrange}
\alias{dc_coordinateuncertainty_inrange}
\title{Data check coordinateuncertainty_inrange Check if coordinateUncertaintyInMeters is within given range.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_coordinateuncertainty_inrange(
  input = NULL,
  lower_limit = 1,
  upper_limit = 20037509
)
}
\arguments{
\item{input}{a vector of coordinate uncertainty in meters. To pass it must
be within lower and upper limits.}

\item{lower_limit}{a numeric value for a lower limit.}

\item{upper_limit}{a numeric value for an upper limit.}
}
\description{
This data check answers: "Is coordinateUncertaintyInMeters between 1 and upper limit?" question.\cr Data check will pass if \strong{The value for coordinateUncertaintyInMeters is within given range.} and will fail if \strong{The value for coordinateUncertaintyInMeters is not within range.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{coordinateUncertaintyInMeters=1}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{coordinateUncertaintyInMeters=-1}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for coordinateUncertaintyInMeters is within given range.
}

\section{sampleFailData}{

The value for coordinateUncertaintyInMeters is not within range.
}

\section{targetDWCField}{

coordinateUncertaintyInMeters
}

\section{checkCategory}{

 location
}

\examples{
perform_dc(data_bats, 'coordinateuncertainty_inrange')
}
\references{
None
}
\keyword{space,conformance,validation}
