% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_countryNameUnknown.R,
%   R/documentation_countryNameUnknown.R
\name{dc_countryNameUnknown}
\alias{dc_countryNameUnknown}
\title{Data check countryNameUnknown}
\format{An object of class \code{"dataCheck"}, see \code{\link{dataCheck}} for details.}
\usage{
dc_countryNameUnknown(TARGET, DEPEND)
}
\description{
This data check answers: "?" question.\cr Data check will pass if \strong{} and will fail if \strong{}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{}, such data check would return \code{}.\cr Example of entries that will fail: \code{}, such data check would return \code{}.
}
\section{samplePassData}{

Country name (dwc:country) is in vocabulary
}

\section{sampleFailData}{

Country name (dwc:country) not in vocabulary
}

\section{targetDWCField}{

country
}

\section{checkCategory}{

 Location
}

\examples{
performDC(DC = dc_countryNameUnknown, DATA = bdchecks::data_bats)
}
\references{
None
}
