% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_datageneralizations_present.R,
%   R/documentation_datageneralizations_present.R
\name{dc_datageneralizations_present}
\alias{dc_datageneralizations_present}
\title{Data check datageneralizations_present Check if dataGeneralizations is not empty.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_datageneralizations_present(TARGET = NULL)
}
\arguments{
\item{TARGET}{a vector of data generalization information.}
}
\description{
This data check answers: "Is dataGeneralizations information present?" question.\cr Data check will pass if \strong{The value for dataGeneralizations is not empty.} and will fail if \strong{The value for dataGeneralizations is not reported.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{dataGeneralizations=record placed on 0.1 degree grid}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{dataGeneralizations=}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for dataGeneralizations is not empty.
}

\section{sampleFailData}{

The value for dataGeneralizations is not reported.
}

\section{targetDWCField}{

dataGeneralizations
}

\section{checkCategory}{

 record-level terms
}

\examples{
perform_dc(data_bats, 'datageneralizations_present')
}
\references{
None
}
\keyword{notificaton,resolution,space,time,name}
