% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_dctype_present.R,
%   R/documentation_dctype_present.R
\name{dc_dctype_present}
\alias{dc_dctype_present}
\title{Data check dctype_present Check if dctype exists and is not empty.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_dctype_present(input = NULL)
}
\arguments{
\item{input}{a vector of dctype information. To pass it must be
non-empty entry.}
}
\description{
This data check answers: "Is dctype information present?"
question.\cr Data check will pass if \strong{The value for dctype is
present.} and will fail if \strong{The value for dctype is missing.}.\cr
Dimension of this data check is \strong{} and it's flagging type is:
\strong{FLAG}\cr Example of entries that will pass: \code{dctype=Image},
such data check would return \code{TRUE}.\cr Example of entries that will
fail: \code{dctype=}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for dctype is present.
}

\section{sampleFailData}{

The value for dctype is missing.
}

\section{targetDWCField}{

type
}

\section{checkCategory}{

 record-level terms
}

\examples{
perform_dc(data_bats, 'dctype_present')
}
\references{
None
}
\keyword{other,validation,completeness,missing,incomplete}
