% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_decimallatitude_inrange.R,
%   R/documentation_decimallatitude_inrange.R
\name{dc_decimallatitude_inrange}
\alias{dc_decimallatitude_inrange}
\title{Data check decimallatitude_inrange Check if decimalLatitude is within parameter range.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_decimallatitude_inrange(input = NULL)
}
\arguments{
\item{input}{a vector of latitude information. To pass it must
be within -90 and 90 degrees.}
}
\description{
This data check answers: "Is decimalLatitude within parameter range?" question.\cr Data check will pass if \strong{The value for decimalLatitude within given numbers.} and will fail if \strong{The value for decimalLatitude is not within given numbers.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{decimalLatitude=0}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{decimalLatitude=91}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for decimalLatitude within given numbers.
}

\section{sampleFailData}{

The value for decimalLatitude is not within given numbers.
}

\section{targetDWCField}{

decimalLatitude
}

\section{checkCategory}{

 location
}

\examples{
perform_dc(data_bats, 'decimallatitude_inrange')
}
\references{
None
}
\keyword{space,validation,conformance}
