% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_decimallongitude_inrange.R,
%   R/documentation_decimallongitude_inrange.R
\name{dc_decimallongitude_inrange}
\alias{dc_decimallongitude_inrange}
\title{Data check decimallongitude_inrange Check if decimalLongitude is within parameter range.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_decimallongitude_inrange(TARGET = NULL)
}
\arguments{
\item{TARGET}{a vector of longitude information. To pass it must
be within -180 and 180 degrees.}
}
\description{
This data check answers: "Is decimalLongitude within parameter range?" question.\cr Data check will pass if \strong{The value for decimalLongitude within given numbers.} and will fail if \strong{The value for decimalLongitude is not within given numbers.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{decimalLongitude=0}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{decimalLongitude=181}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for decimalLongitude within given numbers.
}

\section{sampleFailData}{

The value for decimalLongitude is not within given numbers.
}

\section{targetDWCField}{

decimalLongitude
}

\section{checkCategory}{

 location
}

\examples{
perform_dc(data_bats, 'decimallongitude_inrange')
}
\references{
None
}
\keyword{space,validation,conformance}
