% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_decimallongitude_present.R,
%   R/documentation_decimallongitude_present.R
\name{dc_decimallongitude_present}
\alias{dc_decimallongitude_present}
\title{Data check decimallongitude_present Check if decimalLongitude exists and is not empty.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_decimallongitude_present(TARGET = NULL)
}
\arguments{
\item{TARGET}{a vector of decimallongitude information. To pass it must be
non-empty entry.}
}
\description{
This data check answers: "Is decimalLongitude information present?" question.\cr Data check will pass if \strong{The value for decimalLongitude is present.} and will fail if \strong{The value for decimalLongitude is missing.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{decimalLongitude=-41.0983423}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{decimalLongitude=}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for decimalLongitude is present.
}

\section{sampleFailData}{

The value for decimalLongitude is missing.
}

\section{targetDWCField}{

decimalLongitude
}

\section{checkCategory}{

 location
}

\examples{
perform_dc(data_bats, 'decimallongitude_present')
}
\references{
None
}
\keyword{location,validation,completeness,missing,incomplete}
