% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_depthOutOfRange.R,
%   R/documentation_depthOutOfRange.R
\name{dc_depthOutOfRange}
\alias{dc_depthOutOfRange}
\title{Data check depthOutOfRange}
\format{An object of class function to perform a specific data check.}
\usage{
dc_depthOutOfRange(TARGET, min_dep = 0, max_dep = 11000)
}
\arguments{
\item{TARGET}{a vector of numeric values for depth}

\item{min_dep}{a numeric value specifying minimum depth}

\item{max_dep}{a numeric value specifying maximum depth}
}
\description{
This data check answers: "?" question.\cr Data check will pass if \strong{} and will fail if \strong{}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{}, such data check would return \code{}.\cr Example of entries that will fail: \code{}, such data check would return \code{}.
}
\section{samplePassData}{

Minimum depth is greater than or equal to zero (0) and maximum depth is less than 11,000 meters
}

\section{sampleFailData}{

Minimum depth is less than zero (0) or maximum depth is greater than 11,000 meters
}

\section{targetDWCField}{

minimumDepthInMeters,maximumDepthInMeters
}

\section{checkCategory}{

 Location
}

\examples{
perform_dc(data_bats, 'depthOutOfRange')
}
\references{
None
}
