% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_establishmentmeans_present.R,
%   R/documentation_establishmentmeans_present.R
\name{dc_establishmentmeans_present}
\alias{dc_establishmentmeans_present}
\title{Data check establishmentmeans_present Check if establishmentMeans is not
empty.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_establishmentmeans_present(input = NULL)
}
\arguments{
\item{input}{a vector of establishment means information.}
}
\description{
This data check answers: "Is establishmentMeans information present?"
question.\cr Data check will pass if \strong{The value for
establishmentMeans is not empty.} and will fail if \strong{The value for
establishmentMeans is not reported.}.\cr Dimension of this data check is
\strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries
that will pass: \code{establishmentMeans=cultivated}, such data check would
return \code{TRUE}.\cr Example of entries that will fail:
\code{establishmentMeans=}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for establishmentMeans is not empty.
}

\section{sampleFailData}{

The value for establishmentMeans is not reported.
}

\section{targetDWCField}{

establishmentMeans
}

\section{checkCategory}{

 occurrence
}

\examples{
perform_dc(data_bats, 'establishmentmeans_present')
}
\references{
None
}
\keyword{other,notificaton,completeness}
