% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_eventdate_inrange.R,
%   R/documentation_eventdate_inrange.R
\name{dc_eventdate_inrange}
\alias{dc_eventdate_inrange}
\title{Data check eventdate_inrange Check if eventDate is not empty and is within
given earliest and latest dates.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_eventdate_inrange(
  input = NULL,
  date_start = "1600-01-01",
  date_end = Sys.Date()
)
}
\arguments{
\item{input}{a vector of event date information. To pass must be within 
given dates (start and end).}

\item{date_start}{earliest possible date.}

\item{date_end}{latest possible date.}
}
\description{
This data check answers: "Is eventDate within given dates?"
question.\cr Data check will pass if \strong{The value for eventDate within
given dates.} and will fail if \strong{The value for eventDate is not within
given dates.}.\cr Dimension of this data check is \strong{} and it's
flagging type is: \strong{FLAG}\cr Example of entries that will pass:
\code{eventDate=2001-01-01}, such data check would return \code{TRUE}.\cr
Example of entries that will fail: \code{eventDate=2230-12-31}, such data
check would return \code{FALSE}.
}
\section{samplePassData}{

The value for eventDate within given dates.
}

\section{sampleFailData}{

The value for eventDate is not within given dates.
}

\section{targetDWCField}{

eventDate
}

\section{checkCategory}{

 event
}

\examples{
perform_dc(data_bats, 'eventdate_inrange')
}
\references{
None
}
\keyword{time,validation,conformance}
