% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_eventdate_present.R,
%   R/documentation_eventdate_present.R
\name{dc_eventdate_present}
\alias{dc_eventdate_present}
\title{Data check eventdate_present Check if eventDate exists and is not empty.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_eventdate_present(input = NULL)
}
\arguments{
\item{input}{a vector of eventDate information. To pass it must be
non-empty entry.}
}
\description{
This data check answers: "Is eventDate information present?"
question.\cr Data check will pass if \strong{The value for eventDate is
present.} and will fail if \strong{The value for eventDate is missing.}.\cr
Dimension of this data check is \strong{} and it's flagging type is:
\strong{FLAG}\cr Example of entries that will pass:
\code{eventDate=2001-01-01}, such data check would return \code{TRUE}.\cr
Example of entries that will fail: \code{eventDate=}, such data check would
}
\section{samplePassData}{

The value for eventDate is present.
}

\section{sampleFailData}{

The value for eventDate is missing.
}

\section{targetDWCField}{

eventDate
}

\section{checkCategory}{

 event
}

\examples{
perform_dc(data_bats, 'eventdate_present')
}
\references{
None
}
\keyword{event,time,validation,completeness,missing,incomplete}
