% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_eventDateInFuture.R,
%   R/documentation_eventDateInFuture.R
\name{dc_eventDateInFuture}
\alias{dc_eventDateInFuture}
\title{Data check eventDateInFuture}
\format{An object of class function to perform a specific data check.}
\usage{
dc_eventDateInFuture(TARGET)
}
\description{
This data check answers: "?" question.\cr Data check will pass if \strong{} and will fail if \strong{}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{}, such data check would return \code{}.\cr Example of entries that will fail: \code{}, such data check would return \code{}.
}
\section{samplePassData}{

dwc:eventDate is valid
}

\section{sampleFailData}{

dwc:eventDate is in the future
}

\section{targetDWCField}{

eventDate
}

\section{checkCategory}{

 Event
}

\examples{
performDC(data.checks@dc_body$eventDateInFuture, data_bats)
}
\references{
None
}
