% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_class_export.R
\name{dc_export}
\alias{dc_export}
\title{Export data checks as R objects}
\usage{
dc_export(path_yaml = "./inst/extdata/dataChecks.yaml",
  export_rda = TRUE, export_rox = TRUE, path_rda = "./data/",
  path_rox = "./R/", id_rda = "DC_", id_rox = "DC_")
}
\arguments{
\item{path_yaml}{Path to a YAML file.}

\item{export_rda}{Should function export data check to a .rda file.}

\item{export_rox}{Should function create a documentation in .R for roxygen2.}

\item{path_rda}{Path for .rda file.}

\item{path_rox}{Path for .R (roxygen2) file.}

\item{id_rda}{ID in the beginning of file for .rda file}

\item{id_rox}{ID in the beginning of file for .R file
(helps to differentiate other functions from data checks).}
}
\description{
`dc_export()` is a function for exporting data checks from YAML file to
rda and/or roxygen2 file.
}
