% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_filter.R
\name{dc_filter}
\alias{dc_filter}
\title{Filter Data Checks}
\usage{
dc_filter(dc_result, dc_filts)
}
\arguments{
\item{dc_result}{Object of a dataCheckFlag generated with `perforDataCheck()`}

\item{dc_filts}{A list containing filtering targets and status generated with
`dc_filter_generate()`}
}
\value{
A data.frame that is filtered according to given vector
}
\description{
`dc_filter()` is a function that filters data check result according
 to filtering vector.
}
