% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{dc_filter_generate}
\alias{dc_filter_generate}
\title{Filter Data Checks}
\usage{
dc_filter_generate(
  dc_result_summary,
  cell_selected,
  filters = c("P", "F", "M")
)
}
\arguments{
\item{dc_result_summary}{Summary table for a DataCheckFlagSet class
(must be filterable in `DT`)}

\item{cell_selected}{Cells selected in `DT::datatable`}

\item{filters}{Vector that contains names for passed, failed and missing
data checks}
}
\value{
A list that contains name of a data checks, it's target and
filtering status
}
\description{
`dc_filter_generate()` is a function that generates vector for filtering data
checks result table according to `DT::datatable()` `selectCells` object.
}
