% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{dc_get}
\alias{dc_get}
\title{Load Data Checks}
\usage{
dc_get(path_yaml = "../inst/extdata/dataChecks.yaml")
}
\arguments{
\item{path_yaml}{Path to a YAML file.}
}
\value{
A list of data checks
}
\description{
`dc_get()` is a function for importing data checks from a YAML file as
a dataCheck object.
}
