% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_identifiedDateImprobable.R,
%   R/documentation_identifiedDateImprobable.R
\name{dc_identifiedDateImprobable}
\alias{dc_identifiedDateImprobable}
\title{Data check identifiedDateImprobable}
\format{An object of class function to perform a specific data check.}
\usage{
dc_identifiedDateImprobable(TARGET)
}
\arguments{
\item{TARGET}{a vector to perform data check}
}
\description{
This data check answers: "?" question.\cr Data check will pass if \strong{} and will fail if \strong{}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{}, such data check would return \code{}.\cr Example of entries that will fail: \code{}, such data check would return \code{}.
}
\section{samplePassData}{

The date of identification (dwc:dateIdentified) is post Linnaeus (1753) to the current date
}

\section{sampleFailData}{

The date of identification (dwc:dateIdentified) falls prior to Linnaeus (1753) or after the current date.
}

\section{targetDWCField}{

dateIdentified
}

\section{checkCategory}{

 Identification
}

\examples{
performDC(data.checks@dc_body$identifiedDateImprobable, data_bats)
}
\references{
None
}
