% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_individualCountInvalid.R,
%   R/documentation_individualCountInvalid.R
\name{dc_individualCountInvalid}
\alias{dc_individualCountInvalid}
\title{Data check individualCountInvalid Check if the number of individuals represented present at the time of the occurrence is and intiger}
\format{An object of class function to perform a specific data check.}
\usage{
dc_individualCountInvalid(TARGET)
}
\arguments{
\item{TARGET}{a vector of number of individuals}
}
\description{
This data check answers: "Is number of individuals a positive integer?" question.\cr Data check will pass if \strong{The count of individuals is a valid integer} and will fail if \strong{The count of individuals is not an integer and therefore invalid}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{individualCount=1}, such data check would return \code{Passed}.\cr Example of entries that will fail: \code{individualCount=0.3,individualCount=-1,individualCount=0}, such data check would return \code{Failed}.
}
\section{samplePassData}{

The count of individuals is a valid integer
}

\section{sampleFailData}{

The count of individuals is not an integer and therefore invalid
}

\section{targetDWCField}{

individualCount
}

\section{checkCategory}{

 Occurrence
}

\examples{
perform_dc(data_bats, 'individualCountInvalid')
}
\references{
None
}
