% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_license_present.R,
%   R/documentation_license_present.R
\name{dc_license_present}
\alias{dc_license_present}
\title{Data check license_present Check if license exists and is not empty.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_license_present(TARGET = NULL)
}
\arguments{
\item{TARGET}{a vector of license information. To pass it must be
non-empty entry.}
}
\description{
This data check answers: "Is license information present?" question.\cr Data check will pass if \strong{The value for license is present.} and will fail if \strong{The value for license is missing.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{license=GPL}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{license=}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for license is present.
}

\section{sampleFailData}{

The value for license is missing.
}

\section{targetDWCField}{

license
}

\section{checkCategory}{

 record-level terms
}

\examples{
perform_dc(data_bats, 'license_present')
}
\references{
None
}
\keyword{other,validation,completeness,missing,incomplete}
