% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_export.R
\name{dc_main_create}
\alias{dc_main_create}
\title{Create a data check object}
\usage{
dc_main_create(yaml)
}
\arguments{
\item{yaml}{Data check entry as a list (originally imported as a
YAML file)}
}
\value{
Data check object
}
\description{
Create a data check object from a given YAML file
}
