% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_maxelevation_inrange.R,
%   R/documentation_maxelevation_inrange.R
\name{dc_maxelevation_inrange}
\alias{dc_maxelevation_inrange}
\title{Data check maxelevation_inrange Check if maximumElevationInMeters is
within parameter range.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_maxelevation_inrange(input = NULL, lower_limit = -428, upper_limit = 8850)
}
\arguments{
\item{input}{a vector of maximum elevation meters. To pass it must
be within lower and upper limits.}

\item{lower_limit}{a numeric value for a lower limit.}

\item{upper_limit}{a numeric value for an upper limit.}
}
\description{
This data check answers: "Is maximumElevationInMeters within parameter
range?" question.\cr Data check will pass if \strong{The value for
maximumElevationInMeters within given numbers.} and will fail if \strong{The
value for maximumElevationInMeters is not within given numbers.}.\cr
Dimension of this data check is \strong{} and it's flagging type is:
\strong{FLAG}\cr Example of entries that will pass:
\code{maximumElevationInMeters=0}, such data check would return
\code{TRUE}.\cr Example of entries that will fail:
\code{maximumElevationInMeters=9999}, such data check would return
}
\section{samplePassData}{

The value for maximumElevationInMeters within given numbers.
}

\section{sampleFailData}{

The value for maximumElevationInMeters is not within given numbers.
}

\section{targetDWCField}{

maximumElevationInMeters
}

\section{checkCategory}{

 location
}

\examples{
perform_dc(data_bats, 'maxelevation_inrange')
}
\references{
None
}
\keyword{space,validation,conformance}
