% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_class_create.R
\name{dc_meta_create}
\alias{dc_meta_create}
\title{Create a data check metadata object}
\usage{
dc_meta_create(meta)
}
\arguments{
\item{meta}{Data check metadata entry as a list.}
}
\value{
Data check metadata object
}
\description{
Create a data check metadata object from a given slot in a data check list
object
}
