% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_mindepth_lessthan_maxdepth.R,
%   R/documentation_mindepth_lessthan_maxdepth.R
\name{dc_mindepth_lessthan_maxdepth}
\alias{dc_mindepth_lessthan_maxdepth}
\title{Data check mindepth_lessthan_maxdepth Check if minimumDepthInMeters is not greater than maximumDepthInMeters.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_mindepth_lessthan_maxdepth(input = NULL, input2 = NULL)
}
\arguments{
\item{input}{a vector of minimumDepthInMeters.}

\item{input2}{a vector of maximumDepthInMeters.}
}
\description{
This data check answers: "Is minimumDepthInMeters less or equal to maximumDepthInMeters?" question.\cr Data check will pass if \strong{Given minimumDepthInMeters is not greater than maximumDepthInMeters.} and will fail if \strong{Given minimumDepthInMeters is greater than maximumDepthInMeters.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{minimumDepthInMeters="100",maximumDepthInMeters="200"}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{minimumDepthInMeters="100",maximumDepthInMeters="20"}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

Given minimumDepthInMeters is not greater than maximumDepthInMeters.
}

\section{sampleFailData}{

Given minimumDepthInMeters is greater than maximumDepthInMeters.
}

\section{targetDWCField}{

minimumDepthInMeters
}

\section{checkCategory}{

 location
}

\examples{
perform_dc(data_bats, 'mindepth_lessthan_maxdepth')
}
\references{
None
}
\keyword{space,validation,conformance}
