% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_month_standard.R,
%   R/documentation_month_standard.R
\name{dc_month_standard}
\alias{dc_month_standard}
\title{Data check month_standard Check if month is an integer between 1 and 12 inclusive.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_month_standard(TARGET = NULL)
}
\arguments{
\item{TARGET}{a vector of month information. To pass must be an integer
within 1 and 12.}
}
\description{
This data check answers: "Is month an integer between 1 and 12 inclusive?" question.\cr Data check will pass if \strong{The value for month within given numbers.} and will fail if \strong{The value for month is not within given numbers.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{month=6}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{month=13}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for month within given numbers.
}

\section{sampleFailData}{

The value for month is not within given numbers.
}

\section{targetDWCField}{

month
}

\section{checkCategory}{

 event
}

\examples{
perform_dc(data_bats, 'month_standard')
}
\references{
None
}
\keyword{time,validation,conformance}
