% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_occurrenceid_present.R,
%   R/documentation_occurrenceid_present.R
\name{dc_occurrenceid_present}
\alias{dc_occurrenceid_present}
\title{Data check occurrenceid_present Check if occurrenceID exists and is not empty.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_occurrenceid_present(TARGET = NULL)
}
\arguments{
\item{TARGET}{a vector of occurrence ID information. To pass it must be
non-empty entry.}
}
\description{
This data check answers: "Is occurrenceID information present?" question.\cr Data check will pass if \strong{The value for occurrenceID is present.} and will fail if \strong{The value for occurrenceID is missing.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{occurrenceID=urn:lsid:nhm.ku.edu:Herps:32}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{occurrenceID=}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for occurrenceID is present.
}

\section{sampleFailData}{

The value for occurrenceID is missing.
}

\section{targetDWCField}{

occurrenceID
}

\section{checkCategory}{

 record-level terms
}

\examples{
perform_dc(data_bats, 'occurrenceid_present')
}
\references{
None
}
\keyword{other,validation,completeness,missing,incomplete}
