% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_occurrencestatus_present.R,
%   R/documentation_occurrencestatus_present.R
\name{dc_occurrencestatus_present}
\alias{dc_occurrencestatus_present}
\title{Data check occurrencestatus_present Check if occurrenceStatus exists and
is not empty.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_occurrencestatus_present(input = NULL)
}
\arguments{
\item{input}{a vector of occurrence status information. To pass it must be
non-empty entry.}
}
\description{
This data check answers: "Is occurrenceStatus information present?"
question.\cr Data check will pass if \strong{The value for occurrenceStatus
is present.} and will fail if \strong{The value for occurrenceStatus is
missing.}.\cr Dimension of this data check is \strong{} and it's flagging
type is: \strong{FLAG}\cr Example of entries that will pass:
\code{occurrenceStatus=present}, such data check would return
\code{TRUE}.\cr Example of entries that will fail:
\code{occurrenceStatus=}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for occurrenceStatus is present.
}

\section{sampleFailData}{

The value for occurrenceStatus is missing.
}

\section{targetDWCField}{

occurrenceStatus
}

\section{checkCategory}{

 record-level terms
}

\examples{
perform_dc(data_bats, 'occurrencestatus_present')
}
\references{
None
}
\keyword{other,validation,completeness,missing,incomplete}
