% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_occurrencestatus_standard.R,
%   R/documentation_occurrencestatus_standard.R
\name{dc_occurrencestatus_standard}
\alias{dc_occurrencestatus_standard}
\title{Data check occurrencestatus_standard Check if occurrenceStatus is
specified source authority service.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_occurrencestatus_standard(input = NULL, sources = c("present", "absent"))
}
\arguments{
\item{input}{a vector of occurrenceStatus information. To pass it must be
within given reference dictionary.}

\item{sources}{a vector of available sources.}
}
\description{
This data check answers: "Is occurrenceStatus within given source
authority list?" question.\cr Data check will pass if \strong{The value for
occurrenceStatus is present and correct.} and will fail if \strong{The value
for occurrenceStatus is missing or incorrect.}.\cr Dimension of this data
check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of
entries that will pass: \code{occurrenceStatus=present}, such data check
would return \code{TRUE}.\cr Example of entries that will fail:
\code{occurrenceStatus=observed}, such data check would return
}
\section{samplePassData}{

The value for occurrenceStatus is present and correct.
}

\section{sampleFailData}{

The value for occurrenceStatus is missing or incorrect.
}

\section{targetDWCField}{

occurrenceStatus
}

\section{checkCategory}{

 occurrence
}

\examples{
perform_dc(data_bats, 'occurrencestatus_standard')
}
\references{
None
}
\keyword{vocabulary,validation}
