% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.R
\name{dc_perform}
\alias{dc_perform}
\title{Perform Data Checks}
\usage{
dc_perform(data = NULL, DConly = NULL)
}
\arguments{
\item{data}{Data set to perform data checks}

\item{DConly}{Character vector of names for data checks that should be performed
(ie perform only these data checks)}
}
\value{
Object of a dataCheckFlag class (combined result for all performed
data checks)
}
\description{
`dc_perform()` is a function for performing all available data checks
on a give data set
}
\examples{
dc_perform(data_bats)
}
