% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_phylum_valid.R,
%   R/documentation_phylum_valid.R
\name{dc_phylum_valid}
\alias{dc_phylum_valid}
\title{Data check phylum_valid Check if phylum was found as a value at the rank
of phylum by the specified source authority service.}
\format{An object of class function to perform a specific data check.}
\usage{
dc_phylum_valid(input = NULL)
}
\arguments{
\item{input}{a vector of phylum information. To pass it must be within
gbif phylum ranks.}
}
\description{
This data check answers: "Is phylum information valid?" question.\cr
Data check will pass if \strong{The value for phylum is present and
correct.} and will fail if \strong{The value for phylum is missing or
incorrect.}.\cr Dimension of this data check is \strong{} and it's flagging
type is: \strong{FLAG}\cr Example of entries that will pass:
\code{phylum=Chordate}, such data check would return \code{TRUE}.\cr
Example of entries that will fail: \code{phylum=birds}, such data check
would return \code{FALSE}.
}
\section{samplePassData}{

The value for phylum is present and correct.
}

\section{sampleFailData}{

The value for phylum is missing or incorrect.
}

\section{targetDWCField}{

phylum
}

\section{checkCategory}{

 taxon
}

\examples{
perform_dc(data_bats, 'phylum_valid')
}
\references{
None
}
\keyword{taxon,validation}
