% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_taxo_level.R, R/documentation_taxo_level.R
\name{dc_taxo_level}
\alias{dc_taxo_level}
\title{Data check taxo_level Check if taxonRank is lower than inputted resolution.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_taxo_level(input = NULL, resolution = "species")
}
\arguments{
\item{input}{a vector of taxon rank. To pass must be within given
dictionary.}

\item{resolution}{The low rank of species required}
}
\description{
This data check answers: "Is taxonRank lower than inputted resolution?" question.\cr Data check will pass if \strong{The value for taxonRank is lower than inputted resolution.} and will fail if \strong{The value for taxonRank is higher than inputted resolution.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{taxonRank="species",resolution="family"}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{taxonRank="family",resolution="species"}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for taxonRank is lower than inputted resolution.
}

\section{sampleFailData}{

The value for taxonRank is higher than inputted resolution.
}

\section{targetDWCField}{

taxonRank
}

\section{checkCategory}{

 taxon
}

\examples{
perform_dc(data_bats, 'taxo_level')
}
\references{
None
}
\keyword{taxon,validation,completeness,missing,incomplete}
