% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_taxon_present.R,
%   R/documentation_taxon_present.R
\name{dc_taxon_present}
\alias{dc_taxon_present}
\title{Data check taxon_present Check if at least one field needed to determine the taxon of the entity exists and is not empty.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_taxon_present(TARGET = NULL)
}
\arguments{
\item{TARGET}{a vector of taxon information. To pass it must be non-empty
entry.}
}
\description{
This data check answers: "Is taxon information present?" question.\cr Data check will pass if \strong{The value for at least one term in the Taxon class is present.} and will fail if \strong{The value for all terms organized in the Taxon class are NULL.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{kingdom=animal}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{kingdom=}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for at least one term in the Taxon class is present.
}

\section{sampleFailData}{

The value for all terms organized in the Taxon class are NULL.
}

\section{targetDWCField}{

taxonID, scientificNameID, acceptedNameUsageID, parentNameUsageID, originalNameUsageID, taxonConceptID, scientificName, higherClassification, kingdom, phylum, class, order, family, genus, subgenus, specificEpithet, infraspecificEpithet, vernacularName
}

\section{checkCategory}{

 taxon
}

\examples{
perform_dc(data_bats, 'taxon_present')
}
\references{
None
}
\keyword{taxon,validation,completeness,missing,incomplete}
