% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_taxonrank_standard.R,
%   R/documentation_taxonrank_standard.R
\name{dc_taxonrank_standard}
\alias{dc_taxonrank_standard}
\title{Data check taxonrank_standard Check if taxonRank is standard.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_taxonrank_standard(input = NULL, language = "en", alternative = FALSE)
}
\arguments{
\item{input}{a vector of taxon rank. To pass must be within given
dictionary.}

\item{language}{language for terms (default is English 'en', but usage of
all 'all' is also possible).}

\item{alternative}{logical value - should alternative terms be used too.}
}
\description{
This data check answers: "Is taxonRank information standard?" question.\cr Data check will pass if \strong{The value for taxonRank is standard.} and will fail if \strong{The value for taxonRank is missing or not standard.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{taxonRank=domain}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{taxonRank=realm}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for taxonRank is standard.
}

\section{sampleFailData}{

The value for taxonRank is missing or not standard.
}

\section{targetDWCField}{

taxonRank
}

\section{checkCategory}{

 taxon
}

\examples{
perform_dc(data_bats, 'taxonrank_standard')
}
\references{
None
}
\keyword{taxon,validation,completeness,missing,incomplete}
