% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_validation_continent_empty.R,
%   R/documentation_validation_continent_empty.R
\name{dc_validation_continent_empty}
\alias{dc_validation_continent_empty}
\title{Data check validation_continent_empty Check if continent exists and is not empty.}
\format{An object of class function to perform a specific data check.}
\usage{
dc_validation_continent_empty(TARGET = NULL)
}
\arguments{
\item{TARGET}{a vector of continent information. To pass it must be
non-empty entry.}
}
\description{
This data check answers: "Is continent information present?" question.\cr Data check will pass if \strong{The value for continent is present.} and will fail if \strong{The value for continent is missing.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{continent=Australia}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{continent=}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for continent is present.
}

\section{sampleFailData}{

The value for continent is missing.
}

\section{targetDWCField}{

continent
}

\section{checkCategory}{

 location
}

\examples{
perform_dc(data_bats, 'validation_continent_empty')
}
\references{
None
}
\keyword{location,validation,completeness,missing,incomplete}
