% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_validation_continent_notstandard.R,
%   R/documentation_validation_continent_notstandard.R
\name{dc_validation_continent_notstandard}
\alias{dc_validation_continent_notstandard}
\title{Data check validation_continent_notstandard Check if continent is within a set of available continent names.}
\format{An object of class function to perform a specific data check.}
\usage{
dc_validation_continent_notstandard(TARGET = NULL, sources = c("asia",
  "africa", "europe", "north america", "south america", "australia",
  "oceania", "antarctica", "eurasia", "america"))
}
\arguments{
\item{TARGET}{a vector of continent information. To pass it must be
within given reference dictionary.}

\item{sources}{a vector of available sources.}
}
\description{
This data check answers: "Is continent information valid?" question.\cr Data check will pass if \strong{The value for continent is present and correct.} and will fail if \strong{The value for continent is missing or incorrect.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{continent=Europe}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{continent=North Europe}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for continent is present and correct.
}

\section{sampleFailData}{

The value for continent is missing or incorrect.
}

\section{targetDWCField}{

continent
}

\section{checkCategory}{

 location
}

\examples{
perform_dc(data_bats, 'validation_continent_notstandard')
}
\references{
None
}
\keyword{space,validation,vocabulary,parametrized}
