% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dc_dc_validation_coordinateuncertainty_outofrange.R,
%   R/documentation_validation_coordinateuncertainty_outofrange.R
\name{dc_validation_coordinateuncertainty_outofrange}
\alias{dc_validation_coordinateuncertainty_outofrange}
\title{Data check validation_coordinateuncertainty_outofrange Check if coordinateUncertaintyInMeters is within given range.}
\format{An object of class function to perform a specific data check.}
\usage{
dc_validation_coordinateuncertainty_outofrange(
  TARGET = NULL,
  lower_limit = 1,
  upper_limit = 20037509
)
}
\arguments{
\item{TARGET}{a vector of coordinate uncertainty in meters. To pass it must
be betwee lower and upper limit.}
}
\description{
This data check answers: "Is coordinateUncertaintyInMeters between 1 and upper limit?" question.\cr Data check will pass if \strong{The value for coordinateUncertaintyInMeters is within given range.} and will fail if \strong{The value for coordinateUncertaintyInMeters is not within range.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{coordinateUncertaintyInMeters=1}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{coordinateUncertaintyInMeters=-1}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for coordinateUncertaintyInMeters is within given range.
}

\section{sampleFailData}{

The value for coordinateUncertaintyInMeters is not within range.
}

\section{targetDWCField}{

coordinateUncertaintyInMeters
}

\section{checkCategory}{

 location
}

\examples{
perform_dc(data_bats, 'validation_coordinateuncertainty_outofrange')
}
\references{
None
}
\keyword{space,conformance,validation}
