% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dc_validation_country_countrycode_inconsistent.R,
%   R/documentation_validation_country_countrycode_inconsistent.R
\name{dc_validation_country_countrycode_inconsistent}
\alias{dc_validation_country_countrycode_inconsistent}
\title{Data check validation_country_countrycode_inconsistent Check if country matches country code.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_validation_country_countrycode_inconsistent(TARGET = NULL, TARGET2 = NULL)
}
\arguments{
\item{TARGET}{a vector of country information.
country name (checked against data_countries reference dataset).}

\item{TARGET2}{a vector of country code information.
country name (checked against data_countries reference dataset).}
}
\description{
This data check answers: "Does country matches country code?" question.\cr Data check will pass if \strong{Given country matches country code.} and will fail if \strong{Given country doesn't match country code.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{country="Australia",countryCode="GM"}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{country="Australia",countryCode="AU"}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

Given country matches country code.
}

\section{sampleFailData}{

Given country doesn't match country code.
}

\section{targetDWCField}{

country
}

\section{checkCategory}{

 location
}

\examples{
perform_dc(data_bats, 'validation_country_countrycode_inconsistent')
}
\references{
None
}
\keyword{space,validation,conformance}
