% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_validation_countrycode_empty.R
\name{dc_validation_countrycode_empty}
\alias{dc_validation_countrycode_empty}
\title{Data check validation_countrycode_empty Check if countryCode exists and is not empty.}
\format{
An object of class function to perform a specific data check.
}
\description{
This data check answers: "Is countryCode information present?" question.\cr Data check will pass if \strong{The value for countryCode is present.} and will fail if \strong{The value for countryCode is missing.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{countryCode=AU}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{countryCode=}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for countryCode is present.
}

\section{sampleFailData}{

The value for countryCode is missing.
}

\section{targetDWCField}{

countryCode
}

\section{checkCategory}{

 location
}

\examples{
perform_dc(data_bats, 'validation_countrycode_empty')
}
\references{
None
}
\keyword{location,validation,completeness,missing,incomplete}
