% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_validation_dctype_empty.R,
%   R/documentation_validation_dctype_empty.R
\name{dc_validation_dctype_empty}
\alias{dc_validation_dctype_empty}
\title{Data check validation_dctype_empty Check if dctype exists and is not empty.}
\format{An object of class function to perform a specific data check.}
\usage{
dc_validation_dctype_empty(TARGET = NULL)
}
\arguments{
\item{TARGET}{a vector of dctype information. To pass it must be
non-empty entry.}
}
\description{
This data check answers: "Is dctype information present?" question.\cr Data check will pass if \strong{The value for dctype is present.} and will fail if \strong{The value for dctype is missing.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{dctype=Image}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{dctype=}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for dctype is present.
}

\section{sampleFailData}{

The value for dctype is missing.
}

\section{targetDWCField}{

type
}

\section{checkCategory}{

 Record-level Terms
}

\examples{
perform_dc(data_bats, 'validation_dctype_empty')
}
\references{
None
}
\keyword{other,validation,completness,missing,incomplete}
