% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_validation_genus_notfound.R,
%   R/documentation_validation_genus_notfound.R
\name{dc_validation_genus_notfound}
\alias{dc_validation_genus_notfound}
\title{Data check validation_genus_notfound Check if genus was found as a value at the rank of genus by the specified source authority service.}
\format{An object of class function to perform a specific data check.}
\usage{
dc_validation_genus_notfound(TARGET = NULL)
}
\arguments{
\item{TARGET}{a vector of genus information. To pass it must be within
gbif genus ranks.}
}
\description{
This data check answers: "Is genus information present?" question.\cr Data check will pass if \strong{The value for genus is present and correct.} and will fail if \strong{The value for genus is missing or incorrect.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{genus=Canis}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{genus=}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for genus is present and correct.
}

\section{sampleFailData}{

The value for genus is missing or incorrect.
}

\section{targetDWCField}{

genus
}

\section{checkCategory}{

 taxon
}

\examples{
perform_dc(data_bats, 'validation_genus_notfound')
}
\references{
None
}
\keyword{taxon,validation}
