% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_validation_mindepth_greaterthan_maxdepth.R,
%   R/documentation_validation_mindepth_greaterthan_maxdepth.R
\name{dc_validation_mindepth_greaterthan_maxdepth}
\alias{dc_validation_mindepth_greaterthan_maxdepth}
\title{Data check validation_mindepth_greaterthan_maxdepth Check if minimumDepthInMeters is not greater than maximumDepthInMeters.}
\format{An object of class function to perform a specific data check.}
\usage{
dc_validation_mindepth_greaterthan_maxdepth(TARGET = NULL,
  TARGET2 = NULL)
}
\arguments{
\item{TARGET}{a vector of minimumDepthInMeters.}

\item{TARGET2}{a vector of maximumDepthInMeters.}
}
\description{
This data check answers: "Is minimumDepthInMeters less or equal to maximumDepthInMeters?" question.\cr Data check will pass if \strong{Given minimumDepthInMeters is not greater than maximumDepthInMeters.} and will fail if \strong{Given minimumDepthInMeters is greater than maximumDepthInMeters.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{minimumDepthInMeters="100",maximumDepthInMeters="200"}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{minimumDepthInMeters="100",maximumDepthInMeters="20"}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

Given minimumDepthInMeters is not greater than maximumDepthInMeters.
}

\section{sampleFailData}{

Given minimumDepthInMeters is greater than maximumDepthInMeters.
}

\section{targetDWCField}{

minimumDepthInMeters
}

\section{checkCategory}{

 location
}

\examples{
perform_dc(data_bats, 'validation_mindepth_greaterthan_maxdepth')
}
\references{
None
}
\keyword{space,validation,conformance}
