% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dc_validation_minelevation_greaterthan_maxelevation.R,
%   R/documentation_validation_minelevation_greaterthan_maxelevation.R
\name{dc_validation_minelevation_greaterthan_maxelevation}
\alias{dc_validation_minelevation_greaterthan_maxelevation}
\title{Data check validation_minelevation_greaterthan_maxelevation Check if minimumElevationInMeters is not greater than maximumElevationInMeters.}
\format{An object of class function to perform a specific data check.}
\usage{
dc_validation_minelevation_greaterthan_maxelevation(TARGET = NULL,
  TARGET2 = NULL)
}
\arguments{
\item{TARGET}{a vector of minimumElevationInMeters.}

\item{TARGET2}{a vector of maximumElevationInMeters.}
}
\description{
This data check answers: "Is minimumElevationInMeters less or equal to maximumElevationInMeters?" question.\cr Data check will pass if \strong{Given minimumElevationInMeters is not greater than maximumElevationInMeters.} and will fail if \strong{Given minimumElevationInMeters is greater than maximumElevationInMeters.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{minimumElevationInMeters="100",maximumElevationInMeters="200"}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{minimumElevationInMeters="100",maximumElevationInMeters="20"}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

Given minimumElevationInMeters is not greater than maximumElevationInMeters.
}

\section{sampleFailData}{

Given minimumElevationInMeters is greater than maximumElevationInMeters.
}

\section{targetDWCField}{

minimumElevationInMeters
}

\section{checkCategory}{

 location
}

\examples{
perform_dc(data_bats, 'validation_minelevation_greaterthan_maxelevation')
}
\references{
None
}
\keyword{space,validation,conformance}
