% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_validation_occurrenceid_empty.R,
%   R/documentation_validation_occurrenceid_empty.R
\name{dc_validation_occurrenceid_empty}
\alias{dc_validation_occurrenceid_empty}
\title{Data check validation_occurrenceid_empty Check if occurrenceID exists and is not empty.}
\format{An object of class function to perform a specific data check.}
\usage{
dc_validation_occurrenceid_empty(TARGET = NULL)
}
\arguments{
\item{TARGET}{a vector of occurrence ID information. To pass it must be
non-empty entry.}
}
\description{
This data check answers: "Is occurrenceID information present?" question.\cr Data check will pass if \strong{The value for occurrenceID is present.} and will fail if \strong{The value for occurrenceID is missing.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{occurrenceID=urn:lsid:nhm.ku.edu:Herps:32}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{occurrenceID=}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for occurrenceID is present.
}

\section{sampleFailData}{

The value for occurrenceID is missing.
}

\section{targetDWCField}{

occurrenceID
}

\section{checkCategory}{

 record-level terms
}

\examples{
perform_dc(data_bats, 'validation_occurrenceid_empty')
}
\references{
None
}
\keyword{other,validation,completeness,missing,incomplete}
