% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_validation_occurrencestatus_empty.R,
%   R/documentation_validation_occurrencestatus_empty.R
\name{dc_validation_occurrencestatus_empty}
\alias{dc_validation_occurrencestatus_empty}
\title{Data check validation_occurrencestatus_empty Check if occurrenceStatus exists and is not empty.}
\format{An object of class function to perform a specific data check.}
\usage{
dc_validation_occurrencestatus_empty(TARGET = NULL)
}
\arguments{
\item{TARGET}{a vector of occurrence status information. To pass it must be
non-empty entry.}
}
\description{
This data check answers: "Is occurrenceStatus information present?" question.\cr Data check will pass if \strong{The value for occurrenceStatus is present.} and will fail if \strong{The value for occurrenceStatus is missing.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{occurrenceStatus=present}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{occurrenceStatus=}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for occurrenceStatus is present.
}

\section{sampleFailData}{

The value for occurrenceStatus is missing.
}

\section{targetDWCField}{

occurrenceStatus
}

\section{checkCategory}{

 record-level terms
}

\examples{
perform_dc(data_bats, 'validation_occurrencestatus_empty')
}
\references{
None
}
\keyword{other,validation,completeness,missing,incomplete}
