% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_validation_order_notfound.R,
%   R/documentation_validation_order_notfound.R
\name{dc_validation_order_notfound}
\alias{dc_validation_order_notfound}
\title{Data check validation_order_notfound Check if order was found as a value at the rank of order by the specified source authority service.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_validation_order_notfound(TARGET = NULL)
}
\arguments{
\item{TARGET}{a vector of order information. To pass it must be within
gbif order ranks.}
}
\description{
This data check answers: "Is order information present?" question.\cr Data check will pass if \strong{The value for order is present and correct.} and will fail if \strong{The value for order is missing or incorrect.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{order=Crocodilia}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{order=}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for order is present and correct.
}

\section{sampleFailData}{

The value for order is missing or incorrect.
}

\section{targetDWCField}{

order
}

\section{checkCategory}{

 taxon
}

\examples{
perform_dc(data_bats, 'validation_order_notfound')
}
\references{
None
}
\keyword{taxon,validation}
