% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_validation_phylum_notfound.R,
%   R/documentation_validation_phylum_notfound.R
\name{dc_validation_phylum_notfound}
\alias{dc_validation_phylum_notfound}
\title{Data check validation_phylum_notfound Check if phylum was found as a value at the rank of phylum by the specified source authority service.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_validation_phylum_notfound(TARGET = NULL)
}
\arguments{
\item{TARGET}{a vector of phylum information. To pass it must be within
gbif phylum ranks.}
}
\description{
This data check answers: "Is phylum information present?" question.\cr Data check will pass if \strong{The value for phylum is present and correct.} and will fail if \strong{The value for phylum is missing or incorrect.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{phylum=Chordate}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{phylum=birds}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for phylum is present and correct.
}

\section{sampleFailData}{

The value for phylum is missing or incorrect.
}

\section{targetDWCField}{

phylum
}

\section{checkCategory}{

 taxon
}

\examples{
perform_dc(data_bats, 'validation_phylum_notfound')
}
\references{
None
}
\keyword{taxon,validation}
