% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_validation_scientificname_empty.R,
%   R/documentation_validation_scientificname_empty.R
\name{dc_validation_scientificname_empty}
\alias{dc_validation_scientificname_empty}
\title{Data check validation_scientificname_empty Check if scientificName exists and is not empty.}
\format{An object of class function to perform a specific data check.}
\usage{
dc_validation_scientificname_empty(TARGET = NULL)
}
\arguments{
\item{TARGET}{a vector of scientificname information. To pass it must be
non-empty entry.}
}
\description{
This data check answers: "Is scientificName information present?" question.\cr Data check will pass if \strong{The value for scientificName is present.} and will fail if \strong{The value for scientificName is missing.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{scientificName=banana}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{scientificName=}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for scientificName is present.
}

\section{sampleFailData}{

The value for scientificName is missing.
}

\section{targetDWCField}{

scientificName
}

\section{checkCategory}{

 taxon
}

\examples{
perform_dc(data_bats, 'validation_scientificname_empty')
}
\references{
None
}
\keyword{taxon,validation,completeness,missing,incomplete}
