% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_validation_taxon_empty.R,
%   R/documentation_validation_taxon_empty.R
\name{dc_validation_taxon_empty}
\alias{dc_validation_taxon_empty}
\title{Data check validation_taxon_empty Check if at least one field needed to determine the taxon of the entity exists and is not empty.}
\format{An object of class function to perform a specific data check.}
\usage{
dc_validation_taxon_empty(TARGET)
}
\arguments{
\item{TARGET}{a vector of taxon information. To pass it must be non-empty
entry.}
}
\description{
This data check answers: "Is taxon information present?" question.\cr Data check will pass if \strong{The value for at least one term in the Taxon class is present.} and will fail if \strong{The value for all terms organized in the Taxon class are NULL.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{kingdom=animal}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{kingdom=}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for at least one term in the Taxon class is present.
}

\section{sampleFailData}{

The value for all terms organized in the Taxon class are NULL.
}

\section{targetDWCField}{

taxonID,scientificNameID,acceptedNameUsageID,parentNameUsageID,originalNameUsageID,taxonConceptID,scientificName,higherClassification,kingdom,phylum,class,order,family,genus,subgenus,specificEpithet,infraspecificEpithet,vernacularName
}

\section{checkCategory}{

 taxon
}

\examples{
perform_dc(data_bats, 'validation_taxon_empty')
}
\references{
None
}
\keyword{taxon,validation,completness,missing,incomplete}
