% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_validation_taxonid_empty.R,
%   R/documentation_validation_taxonid_empty.R
\name{dc_validation_taxonid_empty}
\alias{dc_validation_taxonid_empty}
\title{Data check validation_taxonid_empty Check if taxonID exists and is not empty.}
\format{
An object of class function to perform a specific data check.
}
\usage{
dc_validation_taxonid_empty(TARGET = NULL)
}
\arguments{
\item{TARGET}{a vector of taxonid information. To pass it must be non-empty
entry.}
}
\description{
This data check answers: "Is taxonID information present?" question.\cr Data check will pass if \strong{The value for taxonID is present.} and will fail if \strong{The value for taxonID is missing.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{taxonID=101010}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{taxonID=}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for taxonID is present.
}

\section{sampleFailData}{

The value for taxonID is missing.
}

\section{targetDWCField}{

taxonID
}

\section{checkCategory}{

 taxon
}

\examples{
perform_dc(data_bats, 'validation_taxonid_empty')
}
\references{
None
}
\keyword{taxon,validation,completeness,missing,incomplete}
