% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics_methods.R
\docType{methods}
\name{exportDataCheck}
\alias{exportDataCheck}
\alias{exportDataCheck,dataCheckFlag-method}
\title{Export Data Checks}
\usage{
exportDataCheck(object)

\S4method{exportDataCheck}{dataCheckFlag}(object)
}
\arguments{
\item{object}{A result of data checks (data check flag class)}
}
\value{
A data.frame that contains original users data.frame modified according
to data checks
}
\description{
Method that exports data checks
}
