% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCclass_export.R
\name{generate_roxygen_comment}
\alias{generate_roxygen_comment}
\title{Generate roxygen2 documentation from data check object}
\usage{
generate_roxygen_comment(DC)
}
\arguments{
\item{DC}{Data check to generate documentation for.}
}
\value{
Data check description in a roxygen2 comment style
}
\description{
`generate_roxygen_comment()` is a function for generating roxygen2
comments for a given data check. It's not super flexible as it just
inserts metadata into a hard coded skeleton.
}
