% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDataCheck.R
\name{generateDCfilts}
\alias{generateDCfilts}
\title{Filter Data Checks}
\usage{
generateDCfilts(DCresultSummary, selectedCells, filters = c("P", "F",
  "M"))
}
\arguments{
\item{DCresultSummary}{Summary table for a dataCheckFlag class
(must be filterable in `DT`)}

\item{selectedCells}{Cells selected in `DT::datatable`}

\item{filters}{Vector that contains names for passed, failed and missing 
data checks}
}
\value{
A list that contains name of a data checks, it's target and 
filtering status
}
\description{
`generateDCfilts()` is a function that generates vector for filtering data 
checks result table according to `DT::datatable()` `selectedCells` object.
}
