% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{get_github_project}
\alias{get_github_project}
\title{Get Github Project Information}
\usage{
get_github_project(
  repo = "tdwg/bdq",
  project = "Core Tests and Assertions (TG2)",
  header = c(Accept = "application/vnd.github.inertia-preview+json")
)
}
\arguments{
\item{repo}{Repository that we want to work on}

\item{project}{Name of the project that we want to get}

\item{header}{API header (currently it's required for work with projects)}
}
\value{
A data.frame that contains title, url, data information of a given
project
}
\description{
`get_github_project()` is a function for retrieve information for a given github
project
}
