% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_validation_taxonrank_notstandard.R
\name{get_taxon_rank}
\alias{get_taxon_rank}
\title{Retrieve Information about Taxon Rank}
\usage{

  get_taxon_rank(path_rank = "http://rs.gbif.org/vocabulary/gbif/rank.xml")
}
\arguments{
\item{path_rank}{Path to a reference dictionary.}
}
\value{
A data.frame of avaible taxon ranks (prefered and alternative terms)
}
\description{
`get_taxon_rank()` is a function to download reference dictionary of
available taxon ranks.
}
