% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDC.R
\name{getDC}
\alias{getDC}
\title{Load data checks}
\usage{
getDC(pathYAML = "./data/dataChecks.yaml")
}
\arguments{
\item{pathYAML}{Path to a YAML file.}
}
\value{
A list of data checks
}
\description{
`getDC()` is a function for importing data checks from a YAML file as 
a dataCheck object.
}
\examples{
dataChecks <- getDC(pathYAML = "pathToYAML.yaml")

}
