% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DC_identifiedDateImprobable.R
\docType{data}
\name{identifiedDateImprobable}
\alias{identifiedDateImprobable}
\title{Data check identifiedDateImprobable Check if identification date is between Linnaeus and current date}
\format{An object of class \code{"dataCheck"}, see \code{\link{dataCheck}} for details.}
\description{
This data check answers: "Is identification date possible (between Linnaeus and current date)??" question.\cr Data check will pass if \strong{The date of identification (dwc:dateIdentified) is post Linnaeus (1753) to the current date} and will fail if \strong{The date of identification (dwc:dateIdentified) falls prior to Linnaeus (1753) or after the current date.}.\cr Dimension of this data check is \strong{Conformance} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{}, such data check would return \code{}.\cr Example of entries that will fail: \code{}, such data check would return \code{}.
}
\section{samplePassData}{

The date of identification (dwc:dateIdentified) is post Linnaeus (1753) to the current date
}

\section{sampleFailData}{

The date of identification (dwc:dateIdentified) falls prior to Linnaeus (1753) or after the current date.
}

\section{targetDWCField}{

dateIdentified
}

\section{checkCategory}{

 Identification
}

\references{
LATER
}
