% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_datacheck.R
\name{perform_test_dc}
\alias{perform_test_dc}
\title{Perform tests datachecks}
\usage{
perform_test_dc(report = FALSE, wanted_dc = NULL, summary = FALSE)
}
\arguments{
\item{report}{A logical value indicating if html report should be rendered}

\item{wanted_dc}{Character vector of names for data checks that should be 
tested (i.e. test only these data checks)}

\item{summary}{A logical value indicating if the summary should be outputed
instead of the full list of tests}
}
\description{
Performs tests for datachecks and creates report
}
\examples{
data_test <- perform_test_dc()

}
