% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_datacheck.R
\name{perform_testdata_report}
\alias{perform_testdata_report}
\title{Perform tests datachecks and generate report}
\usage{
perform_testdata_report(data_test = NULL)
}
\arguments{
\item{data_test}{List objects of data sets to perform tests}
}
\description{
Perform tests datachecks and generate report
}
