% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performDataCheck.R
\name{performDataCheck}
\alias{performDataCheck}
\title{Perform Data Checks}
\usage{
performDataCheck(data = NULL, DCadd = NULL, DConly = NULL,
  verbose = TRUE, DCstand = ls(pos = ("package:bdchecks"), pattern =
  "^DC_"))
}
\arguments{
\item{data}{Data set to perform data checks}

\item{DCadd}{Character vector of names for additional data checks to be performed}

\item{DConly}{Character vector of names for data checks that should be performed 
(ie perform only these data checks)}

\item{verbose}{Message which data check is being performed}

\item{DCstand}{Character vector of standardize data checks}
}
\value{
Object of a dataCheckFlag class (combined result for all performed
data checks)
}
\description{
`performDataCheck()` is a function for performing all available data checks
on a give data set
}
\examples{
performDataCheck(dataBats)

}
