% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_export.R
\name{roxygen_comment_generate}
\alias{roxygen_comment_generate}
\title{Generate roxygen2 documentation from data check object}
\usage{
roxygen_comment_generate(DC)
}
\arguments{
\item{DC}{Data check to generate documentation for.}
}
\value{
Data check description in a roxygen2 comment style
}
\description{
`roxygen_comment_generate()` is a function for generating roxygen2
comments for a given data check. It's not super flexible as it just
inserts metadata into a hard coded skeleton.
}
