% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runbdchecks.R
\name{run_bdchecks}
\alias{run_bdchecks}
\title{Launch bdchecks Shiny Application}
\usage{
run_bdchecks()
}
\value{
`shiny::runApp()` result within browser.
}
\description{
`bdchecks` is a function that starts biodiversity check `shiny` app.
}
