% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runbdchecks.R
\name{runbdchecks}
\alias{runbdchecks}
\title{Launch bdchecks Shiny Application}
\usage{
runbdchecks()
}
\value{
`shiny::runApp()` result within browser.
}
\description{
`bdchecks` is a function that starts biodiversity check `shiny` app.
}
