% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacheck_helpers.R
\name{tax_available_name}
\alias{tax_available_name}
\title{Return available taxonomical records}
\usage{
tax_available_name(name = NULL, rank_orig = "class", db = "gbif")
}
\arguments{
\item{name}{A character vector of taxonomical records (e.g.
`c("Animalia", "Aves")`).}

\item{rank_orig}{A string for taxonimical rank (e.g. `class`)}

\item{db}{A string for reference database (passed to `tax_available_name`)}
}
\value{
A character vector of available names within given rank.
}
\description{
`tax_available_name()` is a function that checks if given records are within
specified rank. It requires two inputs `name` of taxonimal records and
`rank_orig` (taxonimical rank). Output is subset of names that were within
given rank.
}
\examples{
my_name <- c("Animalia", "Chordata", "Aves", "Mammalia")
my_rank <- "class"
tax_available_name(my_name, my_rank)

}
