% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DC_countryNameUnkown.R
\docType{data}
\name{countryNameUnkown}
\alias{countryNameUnkown}
\title{Data check countryNameUnkown Checks if given country code name is known}
\format{An object of class \code{"dataCheck"}, see \code{\link{dataCheck}} for details.}
\description{
This data check answers: "Is country known?" question.\cr Data check will pass if \strong{Country name (dwc:country) is in vocabulary} and will fail if \strong{Country name (dwc:country) not in vocabulary}.\cr Dimension of this data check is \strong{Conformance} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{AUS,JP}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{xxx, AUST, 1}, such data check would return \code{FALSE}.
}
\references{
LATER
}
\keyword{country,}
\keyword{geo}
