% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDCclass.R
\name{createDCclassMain}
\alias{createDCclassMain}
\title{Create a data check object}
\usage{
createDCclassMain(DCyaml)
}
\arguments{
\item{DCyaml}{Data check entry as a list (originally imported as a 
YAML file)}
}
\value{
Data check object
}
\description{
Create a data check object from a given YAML file
}
