% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDCclass.R
\name{createDCclassMeta}
\alias{createDCclassMeta}
\title{Create a data check metadata object}
\usage{
createDCclassMeta(DCmeta)
}
\arguments{
\item{DCmeta}{Data check metadata entry as a list.}
}
\value{
Data check metadata object
}
\description{
Create a data check metadata object from a given slot in a data check list 
object
}
