% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics_methods.R
\docType{class}
\name{dataCheck-class}
\alias{dataCheck-class}
\alias{dataCheck}
\alias{show,dataCheck-method}
\title{Create Data Check Class}
\usage{
\S4method{show}{dataCheck}(object)
}
\arguments{
\item{object}{a dataCheck object}
}
\description{
Create Data Check Class

Show method for dataCheck objects
}
\section{Slots}{

\describe{
\item{\code{name}}{of a data check}

\item{\code{meta}}{meta-data for a data check of a dataCheckMeta class}

\item{\code{input}}{options for a data check}

\item{\code{func}}{expression to execute}
}}

