% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCclass_create.R, R/summaryDC.R
\docType{class}
\name{dataCheckFlag-class}
\alias{dataCheckFlag-class}
\alias{dataCheckFlag}
\alias{summary}
\alias{summary,dataCheckFlag-method}
\title{Combined data checks class}
\usage{
summary(object, ...)

\S4method{summary}{dataCheckFlag}(object)
}
\description{
Combined data checks class

Method summary
}
\section{Slots}{

\describe{
\item{\code{DC}}{names of performed data checks}

\item{\code{flags}}{list of performed data checks in a dataCheckFlag_SINGLE class}

\item{\code{dataOrig}}{original data before data checks}

\item{\code{dataMod}}{modified data after data checks}
}}

