% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCclass_create.R
\docType{class}
\name{dataCheckFlag-class}
\alias{dataCheckFlag-class}
\alias{dataCheckFlag}
\title{Create data check flag class
For all performed data checks}
\description{
Create data check flag class
For all performed data checks
}
