% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics_methods.R
\docType{class}
\name{dataCheckFlag-class}
\alias{dataCheckFlag-class}
\alias{dataCheckFlag}
\alias{show,dataCheckFlag-method}
\title{Combined Data Checks Class}
\usage{
\S4method{show}{dataCheckFlag}(object)
}
\arguments{
\item{object}{a dataCheckFlag object}
}
\description{
Combined Data Checks Class

Show method for dataCheckFlag objects
}
\section{Slots}{

\describe{
\item{\code{DC}}{names of performed data checks}

\item{\code{flags}}{list of performed data checks in a dataCheckFlag_SINGLE class}

\item{\code{dataOrig}}{original data before data checks}

\item{\code{dataMod}}{modified data after data checks}
}}

