% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCclass_create.R
\docType{class}
\name{dataCheckFlag_SINGLE-class}
\alias{dataCheckFlag_SINGLE-class}
\alias{dataCheckFlag_SINGLE}
\title{Create data check flag class
For a single data check}
\description{
Create data check flag class
For a single data check
}
