% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics_methods.R
\docType{class}
\name{dataCheckFlag_SINGLE-class}
\alias{dataCheckFlag_SINGLE-class}
\alias{dataCheckFlag_SINGLE}
\title{Single Data Check Flag Class}
\description{
Single Data Check Flag Class
}
\section{Slots}{

\describe{
\item{\code{name}}{of performed data check}

\item{\code{target}}{column for performed data check}

\item{\code{flag}}{class}

\item{\code{result}}{logical vector}
}}

