% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCclass_create.R
\docType{class}
\name{dataCheckMeta-class}
\alias{dataCheckMeta-class}
\alias{dataCheckMeta}
\title{Data check meta-data class}
\description{
Data check meta-data class
}
\section{Slots}{

\describe{
\item{\code{description}}{of data check}

\item{\code{keywords}}{for specific data check}

\item{\code{question}}{that this data check answers}

\item{\code{dimension}}{of a data check}

\item{\code{pseudocode}}{for this datachecks}

\item{\code{source}}{creators information}

\item{\code{example}}{of passed and failed inputs}
}}

