% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_generics_methods.R
\docType{class}
\name{dataCheckMeta-class}
\alias{dataCheckMeta-class}
\alias{dataCheckMeta}
\title{Data Check Meta-Data Class}
\description{
Data Check Meta-Data Class
}
\section{Slots}{

\describe{
\item{\code{description}}{of data check}

\item{\code{flags}}{for specific data check}

\item{\code{pseudocode}}{for this datachecks}

\item{\code{source}}{creators information}
}}

