% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DC_dateNull.R
\docType{data}
\name{dateNull}
\alias{dateNull}
\title{Data check dateNull Checks if date feeld is not empty}
\format{An object of class \code{"dataCheck"}, see \code{\link{dataCheck}} for details.}
\description{
This data check answers: "Is date feeld not empty?" question.\cr Data check will pass if \strong{Date information is present} and will fail if \strong{No date information}.\cr Dimension of this data check is \strong{Completness} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{year=2010 + eventDate=2010.01.01 + verbatimEventDate=2010.01.01}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{year=NULL + eventDate=NULL + verbatimEventDate=NULL, year=2010 + eventDate=NULL + verbatimEventDate=NULL}, such data check would return \code{FALSE}.
}
\references{
LATER
}
\keyword{date}
