% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_basisOfRecordBadlyFormed.R,
%   R/documentation_basisOfRecordBadlyFormed.R
\name{dc_basisOfRecordBadlyFormed}
\alias{dc_basisOfRecordBadlyFormed}
\title{Data check basisOfRecordBadlyFormed}
\format{An object of class function to perform a specific data check.}
\usage{
dc_basisOfRecordBadlyFormed(TARGET)
}
\arguments{
\item{TARGET}{a vector to perform data check}
}
\description{
This data check answers: "?" question.\cr Data check will pass if \strong{} and will fail if \strong{}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{}, such data check would return \code{}.\cr Example of entries that will fail: \code{}, such data check would return \code{}.
}
\section{samplePassData}{

dwc:basisOfRecord was matched to vocabulary
}

\section{sampleFailData}{

dwc:basisOfRecord could not be unambiguously matched to vocabulary
}

\section{targetDWCField}{

basisOfRecord
}

\section{checkCategory}{

 Record_level Terms
}

\examples{
perform_dc(data_bats, 'basisOfRecordBadlyFormed')
}
\references{
None
}
