% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_classUnknown.R,
%   R/documentation_classUnknown.R
\name{dc_classUnknown}
\alias{dc_classUnknown}
\title{Data check classUnknown}
\format{An object of class function to perform a specific data check.}
\usage{
dc_classUnknown(TARGET)
}
\arguments{
\item{TARGET}{a vector to perform data check}
}
\description{
This data check answers: "?" question.\cr Data check will pass if \strong{} and will fail if \strong{}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{}, such data check would return \code{}.\cr Example of entries that will fail: \code{}, such data check would return \code{}.
}
\section{samplePassData}{

Taxonomic rank Class was interpreted correctly
}

\section{sampleFailData}{

Taxonomic rank Class value cannot be interpreted
}

\section{targetDWCField}{

class
}

\section{checkCategory}{

 Taxon
}

\examples{
perform_dc(data_bats, 'classUnknown')
}
\references{
None
}
