% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DC_countryMismatch.R
\docType{data}
\name{DC_countryMismatch}
\alias{DC_countryMismatch}
\title{Data check countryMismatch Check if given country match given country code.}
\format{An object of class \code{"dataCheck"}, see \code{\link{dataCheck}} for details.}
\description{
This data check answers: "?" question.\cr Data check will pass if \strong{Country name (dwc:country) and ISO country code (dwc:countryCode) match} and will fail if \strong{Country name (dwc:country) and ISO country code (dwc:countryCode) do not match}.\cr Dimension of this data check is \strong{Consistency} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{}, such data check would return \code{}.\cr Example of entries that will fail: \code{}, such data check would return \code{}.
}
\section{samplePassData}{

Country name (dwc:country) and ISO country code (dwc:countryCode) match
}

\section{sampleFailData}{

Country name (dwc:country) and ISO country code (dwc:countryCode) do not match
}

\section{targetDWCField}{

country,countryCode
}

\section{checkCategory}{

 Location
}

\references{
LATER
}
