% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_countryMismatch.R,
%   R/documentation_countryMismatch.R
\name{dc_countryMismatch}
\alias{dc_countryMismatch}
\title{Data check countryMismatch}
\format{An object of class \code{"dataCheck"}, see \code{\link{dataCheck}} for details.}
\usage{
dc_countryMismatch(TARGET1, TARGET2, DEPEND1, DEPEND2)
}
\description{
This data check answers: "?" question.\cr Data check will pass if \strong{} and will fail if \strong{}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{}, such data check would return \code{}.\cr Example of entries that will fail: \code{}, such data check would return \code{}.
}
\section{samplePassData}{

Country name (dwc:country) and ISO country code (dwc:countryCode) match
}

\section{sampleFailData}{

Country name (dwc:country) and ISO country code (dwc:countryCode) do not match
}

\section{targetDWCField}{

country,countryCode
}

\section{checkCategory}{

 Location
}

\examples{
performDC(DC = dc_countryMismatch, DATA = bdchecks::data_bats)
}
\references{
None
}
