% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_countryNameUnknown.R,
%   R/documentation_countryNameUnknown.R
\name{dc_countryNameUnknown}
\alias{dc_countryNameUnknown}
\title{Data check countryNameUnknown Check if the name of the country or major administrative unit in which the location occurs is within given vocabulary}
\format{An object of class function to perform a specific data check.}
\usage{
dc_countryNameUnknown(TARGET,
  path_reference = system.file("extdata/isocodes.csv", package =
  "bdchecks"))
}
\arguments{
\item{TARGET}{a vector of country names. To pass must be within given
reference set.}

\item{path_reference}{path to reference set of country names.}
}
\description{
This data check answers: "Is country known?" question.\cr Data check will pass if \strong{Country name (dwc:country) is in vocabulary} and will fail if \strong{Country name (dwc:country) not in vocabulary}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{country=Australia}, such data check would return \code{Passed}.\cr Example of entries that will fail: \code{country=Austend}, such data check would return \code{Failed}.
}
\section{samplePassData}{

Country name (dwc:country) is in vocabulary
}

\section{sampleFailData}{

Country name (dwc:country) not in vocabulary
}

\section{targetDWCField}{

country
}

\section{checkCategory}{

 Location
}

\examples{
perform_dc(data_bats, 'countryNameUnknown')
}
\references{
None
}
