% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DC_countryNameUnknown.R
\docType{data}
\name{DC_countryNameUnknown}
\alias{DC_countryNameUnknown}
\title{Data check countryNameUnknown Check if the name of the country or major administrative unit in which the location occurs is within given vocabulary}
\format{An object of class \code{"dataCheck"}, see \code{\link{dataCheck}} for details.}
\description{
This data check answers: "Is country known??" question.\cr Data check will pass if \strong{Country name (dwc:country) is in vocabulary} and will fail if \strong{Country name (dwc:country) not in vocabulary}.\cr Dimension of this data check is \strong{Conformance} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{country=Australia}, such data check would return \code{Passed}.\cr Example of entries that will fail: \code{country=Austend}, such data check would return \code{Failed}.
}
\section{samplePassData}{

Country name (dwc:country) is in vocabulary
}

\section{sampleFailData}{

Country name (dwc:country) not in vocabulary
}

\section{targetDWCField}{

country
}

\section{checkCategory}{

 Location
}

\examples{
performDC(DC = DC_countryNameUnknown, DATA = bdchecks::dataBats)
}
\references{
None
}
