% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_dataGeneralised.R,
%   R/documentation_dataGeneralised.R
\name{dc_dataGeneralised}
\alias{dc_dataGeneralised}
\title{Data check dataGeneralised}
\format{An object of class \code{"dataCheck"}, see \code{\link{dataCheck}} for details.}
\usage{
dc_dataGeneralised()
}
\description{
This data check answers: "?" question.\cr Data check will pass if \strong{} and will fail if \strong{}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{}, such data check would return \code{}.\cr Example of entries that will fail: \code{}, such data check would return \code{}.
}
\section{samplePassData}{

Latitude and longitude values have not been generalized as indicated by dwc:dataGeneralizations is NULL
}

\section{sampleFailData}{

Latitude and longitude values may have been generalized as indicated by dwc:dataGeneralizations is not NULL
}

\section{targetDWCField}{

dataGeneralizations
}

\section{checkCategory}{

 Record_level Terms
}

\examples{
performDC(DC = dc_dataGeneralised, DATA = bdchecks::data_bats)
}
\references{
None
}
