% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_dateIdentifiedInFuture.R,
%   R/documentation_dateIdentifiedInFuture.R
\name{dc_dateIdentifiedInFuture}
\alias{dc_dateIdentifiedInFuture}
\title{Data check dateIdentifiedInFuture}
\format{An object of class function to perform a specific data check.}
\usage{
dc_dateIdentifiedInFuture(TARGET)
}
\arguments{
\item{TARGET}{a vector to perform data check}
}
\description{
This data check answers: "?" question.\cr Data check will pass if \strong{} and will fail if \strong{}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{}, such data check would return \code{}.\cr Example of entries that will fail: \code{}, such data check would return \code{}.
}
\section{samplePassData}{

Date (dwc:dateIdentified) is not in the future
}

\section{sampleFailData}{

Date (dwc:dateIdentified) is in the future
}

\section{targetDWCField}{

dateIdentified
}

\section{checkCategory}{

 Identification
}

\examples{
performDC(data.checks@dc_body$dateIdentifiedInFuture, data_bats)
}
\references{
None
}
