% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_dateNull.R, R/documentation_dateNull.R
\name{dc_dateNull}
\alias{dc_dateNull}
\title{Data check dateNull}
\format{An object of class function to perform a specific data check.}
\usage{
dc_dateNull(TARGET, date_format = "\%Y-\%m-\%d")
}
\arguments{
\item{TARGET}{a vector of event date information}

\item{date_format}{a string for full date format}
}
\description{
This data check answers: "?" question.\cr Data check will pass if \strong{} and will fail if \strong{}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{}, such data check would return \code{}.\cr Example of entries that will fail: \code{}, such data check would return \code{}.
}
\section{samplePassData}{

Date information is present
}

\section{sampleFailData}{

No date information
}

\section{targetDWCField}{

eventDate,year,verbatimEventDate
}

\section{checkCategory}{

 Event
}

\examples{
perform_dc(data_bats, 'dateNull')
}
\references{
None
}
