% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_earliest_date.R,
%   R/documentation_earliest_date.R
\name{dc_earliest_date}
\alias{dc_earliest_date}
\title{Data check earliest_date Check if eventDate is greater than inputted
resolution.}
\format{An object of class function to perform a specific data check.}
\usage{
dc_earliest_date(input = NULL, lowest_date = "2018-02-02", ...)
}
\arguments{
\item{input}{a vector of event date information. To pass must be greater 
than provided resolution.}

\item{lowest_date}{a date for the lowest date required.}

\item{...}{to allow other input-based to run simultaneously.}
}
\description{
This data check answers: "Is eventDate greater than inputted
resolution?" question.\cr Data check will pass if \strong{The value for
eventDate is greater than provided resolution.} and will fail if \strong{The
value for eventDate is not greater than provided resolution.}.\cr Dimension
of this data check is \strong{} and it's flagging type is:
\strong{FLAG}\cr Example of entries that will pass:
\code{eventDate="2002-02-02",provided_input="1996-07-06"}, such data check
would return \code{TRUE}.\cr Example of entries that will fail:
\code{eventDate="2002-02-02",provided_input="2020-06-01"}, such data check
}
\section{samplePassData}{

The value for eventDate is greater than provided resolution.
}

\section{sampleFailData}{

The value for eventDate is not greater than provided resolution.
}

\section{targetDWCField}{

eventDate
}

\section{checkCategory}{

 event
}

\examples{
perform_dc(
  data_bats, 'earliest_date', flags = FALSE, c("2018-02-02")
)
}
\references{
None
}
\keyword{time,validation,conformance}
