% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_elevationOutOfRange.R,
%   R/documentation_elevationOutOfRange.R
\name{dc_elevationOutOfRange}
\alias{dc_elevationOutOfRange}
\title{Data check elevationOutOfRange}
\format{An object of class function to perform a specific data check.}
\usage{
dc_elevationOutOfRange(TARGET, min_ele = 0, max_ele = 10000)
}
\arguments{
\item{TARGET}{a vector of numeric values for elevation}

\item{min_ele}{a numeric value specifying minimum elevation}

\item{max_ele}{a numeric value specifying maximum elevation}
}
\description{
This data check answers: "?" question.\cr Data check will pass if \strong{} and will fail if \strong{}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{}, such data check would return \code{}.\cr Example of entries that will fail: \code{}, such data check would return \code{}.
}
\section{samplePassData}{

Minimum elevation in meters is greater or equal to zero and maximum elevation in meters is less than 10,000 meters
}

\section{sampleFailData}{

Minimum elevation in meters is less than zero and/or maximum elevation in meters is greater than 10,000
}

\section{targetDWCField}{

minimumElevationinMeters,maximumElevationInMeters
}

\section{checkCategory}{

 Location
}

\examples{
perform_dc(data_bats, 'elevationOutOfRange')
}
\references{
None
}
