% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DC_eventDateInFuture.R
\docType{data}
\name{DC_eventDateInFuture}
\alias{DC_eventDateInFuture}
\title{Data check eventDateInFuture Check if event date is not in the future}
\format{An object of class \code{"dataCheck"}, see \code{\link{dataCheck}} for details.}
\description{
This data check answers: "?" question.\cr Data check will pass if \strong{dwc:eventDate is valid} and will fail if \strong{dwc:eventDate is in the future}.\cr Dimension of this data check is \strong{Conformance} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{eventDate=2001-01-01}, such data check would return \code{Passed}.\cr Example of entries that will fail: \code{eventDate=2230-12-31}, such data check would return \code{Failed}.
}
\section{samplePassData}{

dwc:eventDate is valid
}

\section{sampleFailData}{

dwc:eventDate is in the future
}

\section{targetDWCField}{

eventDate
}

\section{checkCategory}{

 Event
}

\examples{
performDC(DC = DC_eventDateInFuture, DATA = bdchecks::dataBats)
}
\references{
None
}
