% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_identifiedDateImprobable.R,
%   R/documentation_identifiedDateImprobable.R
\name{dc_identifiedDateImprobable}
\alias{dc_identifiedDateImprobable}
\title{Data check identifiedDateImprobable Check if identification date is between Linnaeus and current date}
\format{An object of class function to perform a specific data check.}
\usage{
dc_identifiedDateImprobable(TARGET, date_Linnaeus = "1753-01-01",
  date_format = "\%Y-\%m-\%d")
}
\arguments{
\item{TARGET}{a vector of identification date. To pass must be a valid
format between Linnaeus and present.}

\item{date_Linnaeus}{a date for Linnaeus to start counting from.}

\item{date_format}{a string for full date format.}
}
\description{
This data check answers: "Is identification date possible (between Linnaeus and current date)?" question.\cr Data check will pass if \strong{The date of identification (dwc:dateIdentified) is post Linnaeus (1753) to the current date} and will fail if \strong{The date of identification (dwc:dateIdentified) falls prior to Linnaeus (1753) or after the current date.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{dateidentified=2000-01-01}, such data check would return \code{Passed}.\cr Example of entries that will fail: \code{dateidentified=1573-02-14,dateidentified=1000-01-01,dateidentified=3000-01-01}, such data check would return \code{Failed}.
}
\section{samplePassData}{

The date of identification (dwc:dateIdentified) is post Linnaeus (1753) to the current date
}

\section{sampleFailData}{

The date of identification (dwc:dateIdentified) falls prior to Linnaeus (1753) or after the current date.
}

\section{targetDWCField}{

dateIdentified
}

\section{checkCategory}{

 Identification
}

\examples{
perform_dc(data_bats, 'identifiedDateImprobable')
}
\references{
None
}
