% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_individualcountInvalid.R,
%   R/documentation_individualcountInvalid.R
\name{dc_individualcountInvalid}
\alias{dc_individualcountInvalid}
\title{Data check individualcountInvalid}
\format{An object of class function to perform a specific data check.}
\usage{
dc_individualcountInvalid(TARGET)
}
\description{
This data check answers: "?" question.\cr Data check will pass if \strong{} and will fail if \strong{}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{}, such data check would return \code{}.\cr Example of entries that will fail: \code{}, such data check would return \code{}.
}
\section{samplePassData}{

The count of individuals is a valid integer
}

\section{sampleFailData}{

The count of individuals is not an integer and therefore invalid
}

\section{targetDWCField}{

individualCount
}

\section{checkCategory}{

 Occurrence
}

\examples{
performDC(data.checks@dc_body$individualcountInvalid, data_bats)
}
\references{
None
}
