% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DC_individualcountInvalid.R
\docType{data}
\name{DC_individualcountInvalid}
\alias{DC_individualcountInvalid}
\title{Data check individualcountInvalid Check if the number of individuals represented present at the time of the occurrence is and intiger}
\format{An object of class \code{"dataCheck"}, see \code{\link{dataCheck}} for details.}
\description{
This data check answers: "?" question.\cr Data check will pass if \strong{The count of individuals is a valid integer} and will fail if \strong{The count of individuals is not an integer and therefore invalid}.\cr Dimension of this data check is \strong{Conformance} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{individualCount=1}, such data check would return \code{Passed}.\cr Example of entries that will fail: \code{individualCount=0.3,individualCount=-1,individualCount=0}, such data check would return \code{Failed}.
}
\section{samplePassData}{

The count of individuals is a valid integer
}

\section{sampleFailData}{

The count of individuals is not an integer and therefore invalid
}

\section{targetDWCField}{

individualCount
}

\section{checkCategory}{

 Occurrence
}

\examples{
performDC(DC = DC_individualcountInvalid, DATA = bdchecks::dataBats)
}
\references{
None
}
