% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_modifiedInFuture.R,
%   R/documentation_modifiedInFuture.R
\name{dc_modifiedInFuture}
\alias{dc_modifiedInFuture}
\title{Data check modifiedInFuture}
\format{An object of class \code{"dataCheck"}, see \code{\link{dataCheck}} for details.}
\usage{
dc_modifiedInFuture(TARGET)
}
\description{
This data check answers: "?" question.\cr Data check will pass if \strong{} and will fail if \strong{}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{}, such data check would return \code{}.\cr Example of entries that will fail: \code{}, such data check would return \code{}.
}
\section{samplePassData}{

dcterms:modified date is valid
}

\section{sampleFailData}{

dcterms:modified is in the future
}

\section{targetDWCField}{

modified
}

\section{checkCategory}{

 Record_level Terms
}

\examples{
performDC(DC = dc_modifiedInFuture, DATA = bdchecks::data_bats)
}
\references{
None
}
