% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_monthInvalid.R,
%   R/documentation_monthInvalid.R
\name{dc_monthInvalid}
\alias{dc_monthInvalid}
\title{Data check monthInvalid}
\format{An object of class \code{"dataCheck"}, see \code{\link{dataCheck}} for details.}
\usage{
dc_monthInvalid(TARGET)
}
\description{
This data check answers: "?" question.\cr Data check will pass if \strong{} and will fail if \strong{}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{}, such data check would return \code{}.\cr Example of entries that will fail: \code{}, such data check would return \code{}.
}
\section{samplePassData}{

The event month is between 1 and 12
}

\section{sampleFailData}{

The event month is less than 1 or is greater than 12
}

\section{targetDWCField}{

month
}

\section{checkCategory}{

 Event
}

\examples{
performDC(DC = dc_monthInvalid, DATA = bdchecks::data_bats)
}
\references{
None
}
