% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_occurrenceIdNotGuid.R,
%   R/documentation_occurrenceIdNotGuid.R
\name{dc_occurrenceIdNotGuid}
\alias{dc_occurrenceIdNotGuid}
\title{Data check occurrenceIdNotGuid}
\format{An object of class \code{"dataCheck"}, see \code{\link{dataCheck}} for details.}
\usage{
dc_occurrenceIdNotGuid(TARGET)
}
\description{
This data check answers: "?" question.\cr Data check will pass if \strong{} and will fail if \strong{}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{}, such data check would return \code{}.\cr Example of entries that will fail: \code{}, such data check would return \code{}.
}
\section{samplePassData}{

occurrenceID is a globally uninique identifier (GUID)
}

\section{sampleFailData}{

occurrenceID is an integer, assuring that it is not a globally uninique identifier (GUID)
}

\section{targetDWCField}{

occurrenceID
}

\section{checkCategory}{

 Occurrence
}

\examples{
performDC(DC = dc_occurrenceIdNotGuid, DATA = bdchecks::data_bats)
}
\references{
None
}
