% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DC_occurrenceIdNotGuid.R
\docType{data}
\name{DC_occurrenceIdNotGuid}
\alias{DC_occurrenceIdNotGuid}
\title{Data check occurrenceIdNotGuid Check if an identifier for the occurrence is a globally uniquely identifier (currently we use regex pattern solution is implemented)}
\format{An object of class \code{"dataCheck"}, see \code{\link{dataCheck}} for details.}
\description{
This data check answers: "?" question.\cr Data check will pass if \strong{occurrenceID is a globally uninique identifier (GUID)} and will fail if \strong{occurrenceID is an integer, assuring that it is not a globally uninique identifier (GUID)}.\cr Dimension of this data check is \strong{Conformance} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{3cfe9ab4-79f8-4afd-8da5-723183ef16a3}, such data check would return \code{Passed}.\cr Example of entries that will fail: \code{occurrenceID=42}, such data check would return \code{Failed}.
}
\section{samplePassData}{

occurrenceID is a globally uninique identifier (GUID)
}

\section{sampleFailData}{

occurrenceID is an integer, assuring that it is not a globally uninique identifier (GUID)
}

\section{targetDWCField}{

occurrenceID
}

\section{checkCategory}{

 Occurrence
}

\examples{
performDC(DC = DC_occurrenceIdNotGuid, DATA = bdchecks::dataBats)
}
\references{
None
}
