% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_precisionRangeMismatch.R,
%   R/documentation_precisionRangeMismatch.R
\name{dc_precisionRangeMismatch}
\alias{dc_precisionRangeMismatch}
\title{Data check precisionRangeMismatch}
\format{An object of class function to perform a specific data check.}
\usage{
dc_precisionRangeMismatch(TARGET)
}
\arguments{
\item{TARGET}{a vector to perform data check}
}
\description{
This data check answers: "?" question.\cr Data check will pass if \strong{} and will fail if \strong{}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{}, such data check would return \code{}.\cr Example of entries that will fail: \code{}, such data check would return \code{}.
}
\section{samplePassData}{

The coordinate precision (dwc:coordinatePrecision) is between zero (minimum) and one (maximum)
}

\section{sampleFailData}{

The coordinate precision (dwc:coordinatePrecision), as a decimal representation, is outside the range of zero (minimum) and one (maximum)
}

\section{targetDWCField}{

coordinatePrecision
}

\section{checkCategory}{

 Location
}

\examples{
performDC(data.checks@dc_body$precisionRangeMismatch, data_bats)
}
\references{
None
}
