% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_uncertaintyRangeMismatch.R,
%   R/documentation_uncertaintyRangeMismatch.R
\name{dc_uncertaintyRangeMismatch}
\alias{dc_uncertaintyRangeMismatch}
\title{Data check uncertaintyRangeMismatch}
\format{An object of class function to perform a specific data check.}
\usage{
dc_uncertaintyRangeMismatch(TARGET)
}
\arguments{
\item{TARGET}{a vector to perform data check}
}
\description{
This data check answers: "?" question.\cr Data check will pass if \strong{} and will fail if \strong{}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{}, such data check would return \code{}.\cr Example of entries that will fail: \code{}, such data check would return \code{}.
}
\section{samplePassData}{

Geopoint uncertainty (dwc:coordinateUncertaintyInMeters) is a whole number and greater than zero (meters)
}

\section{sampleFailData}{

Geopoint uncertainty (dwc:coordinateUncertaintyInMeters) should be a whole number and greater than zero (meters)
}

\section{targetDWCField}{

coordinateUncertaintyInMeters
}

\section{checkCategory}{

 Location
}

\examples{
performDC(data.checks@dc_body$uncertaintyRangeMismatch, data_bats)
}
\references{
None
}
