% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/documentation_validation_taxonrank_notstandard.R
\name{dc_validation_taxonrank_notstandard}
\alias{dc_validation_taxonrank_notstandard}
\title{Data check validation_taxonrank_notstandard Check if taxonRank is standard.}
\format{
An object of class function to perform a specific data check.
}
\description{
This data check answers: "Is taxonRank information standard?" question.\cr Data check will pass if \strong{The value for taxonRank is standard.} and will fail if \strong{The value for taxonRank is missing or not standard.}.\cr Dimension of this data check is \strong{} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{taxonRank=domain}, such data check would return \code{TRUE}.\cr Example of entries that will fail: \code{taxonRank=realm}, such data check would return \code{FALSE}.
}
\section{samplePassData}{

The value for taxonRank is standard.
}

\section{sampleFailData}{

The value for taxonRank is missing or not standard.
}

\section{targetDWCField}{

taxonRank
}

\section{checkCategory}{

 taxon
}

\examples{
perform_dc(data_bats, 'validation_taxonrank_notstandard')
}
\references{
None
}
\keyword{taxon,validation,completeness,missing,incomplete}
