% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DC_yearMissing.R
\docType{data}
\name{DC_yearMissing}
\alias{DC_yearMissing}
\title{Data check yearMissing Check if year information is not missing}
\format{An object of class \code{"dataCheck"}, see \code{\link{dataCheck}} for details.}
\description{
This data check answers: "Is year information present??" question.\cr Data check will pass if \strong{The value for dwc:year is valid} and will fail if \strong{The value for dwc:year is NULL}.\cr Dimension of this data check is \strong{Completeness} and it's flagging type is: \strong{FLAG}\cr Example of entries that will pass: \code{year=2000}, such data check would return \code{Passed}.\cr Example of entries that will fail: \code{year=}, such data check would return \code{Faield}.
}
\section{samplePassData}{

The value for dwc:year is valid
}

\section{sampleFailData}{

The value for dwc:year is NULL
}

\section{targetDWCField}{

year
}

\section{checkCategory}{

 Event
}

\examples{
performDC(DC = DC_yearMissing, DATA = bdchecks::dataBats)
}
\references{
None
}
