% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCclass_export.R
\name{exportDC}
\alias{exportDC}
\title{Export data checks as R objects}
\usage{
exportDC(pathYAML = "./inst/extdata/dataChecks.yaml", exportRDA = TRUE,
  exportROX = TRUE, pathRDA = "./data/", pathROX = "./R/",
  idRDA = "DC_", idROX = "DC_")
}
\arguments{
\item{pathYAML}{Path to a YAML file.}

\item{exportRDA}{Should function export data check to a .rda file.}

\item{exportROX}{Should function create a documentation in .R for roxygen2.}

\item{pathRDA}{Path for .rda file.}

\item{pathROX}{Path for .R (roxygen2) file.}

\item{idRDA}{ID in the beginning of file for .rda file}

\item{idROX}{ID in the beginning of file for .R file 
(helps to differentiate other functions from data checks).}
}
\description{
`exportDC()` is a function for exporting data checks from YAML file to
rda and/or roxygen2 file.
}
