% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.R
\name{perform_dc_missing}
\alias{perform_dc_missing}
\title{Find Missing Values}
\usage{
perform_dc_missing(result, input)
}
\arguments{
\item{result}{Data check modified input vector}

\item{input}{Original input column that's submitted to datacheck}
}
\value{
result vector with inserted NA and FALSE values
}
\description{
Internal function that deals with missing values. For original missing 
values ("" or NA) it returns NA. For generated NA values it returns NA.
}
