% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_datacheck.R
\name{perform_test_dc}
\alias{perform_test_dc}
\title{Perform tests datachecks}
\usage{
perform_test_dc(report = FALSE)
}
\arguments{
\item{report}{A logical value indicating if html report should be rendered}
}
\description{
Performs tests for datachecks and creates report
}
\examples{
data_test <- perform_test_dc()

}
