% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_datacheck.R
\name{perform_testdata}
\alias{perform_testdata}
\title{Perform tests datachecks}
\usage{
perform_testdata(report = FALSE)
}
\arguments{
\item{report}{A logical value idicating if html report should be rendered}
}
\description{
Perform tests datachecks
}
