% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryDC.R
\name{summary_DC}
\alias{summary_DC}
\title{Create a data check metatadata from a given YAML file}
\usage{
summary_DC(object, fancy = TRUE)
}
\arguments{
\item{fancy}{Nice kable outptu}
}
\value{
result
}
\description{
Create a data check metatadata from a given YAML file
}
