% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_dc}
\alias{summary_dc}
\title{Summarise Data Checks}
\usage{
summary_dc(input_flag, fancy = TRUE, filtering_dt = FALSE)
}
\arguments{
\item{input_flag}{Object of a DataCheckFlagSet class}

\item{fancy}{Should ouput be returned in a rst format}

\item{filtering_dt}{Should output be returned as a summary table that could
be parsed with a `DT` pacakge}
}
\value{
A data.frame or rst table with summary statistics
}
\description{
`summary_dc()` is a function that calculated statistics for how many data
checks passed. It's main input is an object of a DataCheckFlagSet class and output
is a summary table.
}
\examples{
result <- perform_dc(data_bats)
# Fancy summary table (for usage in reports)
summary_dc(result)
# object of class used for data filtering data.frame
summary_dc(result, fancy = FALSE, filtering_dt = TRUE)

}
